% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{approx_cpptsemOptimx}
\alias{approx_cpptsemOptimx}
\title{approx_cpptsemOptimx}
\usage{
approx_cpptsemOptimx(
  cpptsemmodel,
  regM2LLCpptsem,
  gradCpptsem,
  startingValues,
  adaptiveLassoWeights,
  N,
  lambda,
  regIndicators,
  targetVector,
  epsilon,
  objective,
  testGradients,
  controlOptimx,
  silent
)
}
\arguments{
\item{cpptsemmodel}{model returned from cpptsem}

\item{regM2LLCpptsem}{regularized fitting function}

\item{gradCpptsem}{function for computing the gradients of regM2LLCpptsem}

\item{startingValues}{starting values for optimization}

\item{adaptiveLassoWeights}{vector with weights of the adaptive lasso}

\item{N}{sample size}

\item{lambda}{tuning parameter lambda}

\item{regIndicators}{string vector with names of regularized parameters}

\item{targetVector}{named vector with values towards which the parameters are regularized}

\item{epsilon}{tuning parameter for epsL1 approximation}

\item{objective}{ML or Kalman}

\item{testGradients}{should be tested if the final parameters result in NA gradients?}

\item{controlOptimx}{additional arguments passed as control to optimx}

\item{silent}{suppress warnings}
}
\description{
creates an approximate solution to regularized ctsem using optimx
}
\author{
Jannik Orzek
}
