% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_optimization.R
\name{approx_createFitFunString}
\alias{approx_createFitFunString}
\title{approx_createFitFunString}
\usage{
approx_createFitFunString(
  penalty,
  elastic_alpha = NULL,
  elastic_gamma = NULL,
  epsilon = 1e-05,
  silent = FALSE
)
}
\arguments{
\item{penalty}{type of penalty. Currently supported are lasso and ridge}

\item{elastic_alpha}{PLACEHOLDER. Used in elastic net. Not yet implemented}

\item{elastic_gamma}{PLACEHOLDER. Used in elastic net. Not yet implemented}

\item{epsilon}{epsilon is used to transform the non-differentiable lasso penalty to a differentible one}

\item{silent}{Boolean indicating if print statements will be suppressed}
}
\description{
Creates the new approx_createFitFunString for specified regularization
}
\details{
NOTE: Function located in file approx_optimization.R
}
\author{
Jannik Orzek
}
\keyword{internal}
