% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_optimization.R
\name{approx_getCVFit}
\alias{approx_getCVFit}
\title{approx_getCVFit}
\usage{
approx_getCVFit(
  mxObject,
  ctsemObject,
  cvSample,
  objective,
  fitAndParameters,
  parameterLabels,
  lambdas
)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{ctsemObject}{Fitted object of class ctsemOMX}

\item{cvSample}{only required if objective = "Kalman" and ctsemObject ist of type ctsemInit. Please provide a data set in wide format compatible to ctsemOMX}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{fitAndParameters}{table with fit results and parameter values}

\item{parameterLabels}{vector with parameter labels}

\item{lambdas}{vector with lambdas}
}
\description{
computes cv fit for approximate optimization
}
\details{
NOTE: Function located in file approx_optimization.R
}
\author{
Jannik Orzek
}
