% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_optimization.R
\name{approx_getFitIndices}
\alias{approx_getFitIndices}
\title{approx_getFitIndices}
\usage{
approx_getFitIndices(
  fittedRegModel,
  sampleSize,
  regOn = "DRIFT",
  regIndicators,
  returnFitIndices = TRUE,
  cvSample = NULL,
  zeroThresh = 0.001
)
}
\arguments{
\item{fittedRegModel}{fitted regCtsem object}

\item{sampleSize}{sample size}

\item{regOn}{string specifying which matrix should be regularized. Currently only supports DRIFT}

\item{regIndicators}{matrix with ones and zeros specifying which parameters in regOn should be regularized. Must be of same size as the regularized matrix. 1 = regularized, 0 = not regularized. Alternatively, labels for the regularized parameters can be used (e.g. drift_eta1_eta2)}

\item{returnFitIndices}{Boolean: should fit indices be returned?}

\item{cvSample}{cross-validation sample. Has to be of type mxData}

\item{zeroThresh}{threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx}
}
\description{
computes fit indices
}
\details{
NOTE: Function located in file approx_optimization.R
}
\author{
Jannik Orzek
}
