% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_gradCpptsem.R
\name{approx_gradCpptsem}
\alias{approx_gradCpptsem}
\title{approx_gradCpptsem}
\usage{
approx_gradCpptsem(
  parameters,
  cpptsemmodel,
  adaptiveLassoWeights,
  N,
  lambda,
  regIndicators,
  epsilon,
  objective,
  failureReturns
)
}
\arguments{
\item{parameters}{parameter values}

\item{adaptiveLassoWeights}{vector with weights of the adaptive lasso}

\item{N}{sample size}

\item{lambda}{tuning parameter lambda}

\item{regIndicators}{string vector with names of regularized parameters}

\item{epsilon}{tuning parameter for epsL1 approximation}

\item{objective}{ML or Kalman}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}

\item{maxit}{maximal number of iterations}

\item{testGradients}{should be tested if the final parameters result in NA gradients?}
}
\description{
computes gradients for an approximate optimization of regularized ctsem based on cpptsem
}
\author{
Jannik Orzek
}
