% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_initializeModel.R
\name{approx_initializeModel}
\alias{approx_initializeModel}
\title{approx_initializeModel}
\usage{
approx_initializeModel(
  ctsemObject = NULL,
  mxObject = NULL,
  sampleSize = NULL,
  regOn = "DRIFT",
  regIndicators,
  regValue,
  penalty = "lasso",
  elastic_alpha = NULL,
  elastic_gamma = NULL,
  adaptiveLassoWeights = NULL,
  returnFitIndices = TRUE,
  cvSample = NULL,
  autoCV = FALSE,
  k = 5,
  epsilon = 0.001,
  zeroThresh = 0.001,
  cores = 1,
  verbose = 0,
  silent = FALSE,
  progressBar = TRUE,
  parallelProgressBar = NULL
)
}
\arguments{
\item{ctsemObject}{Fitted object of class ctsemOMX}

\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{sampleSize}{sample size}

\item{regOn}{string specifying which matrix should be regularized. Currently only supports DRIFT}

\item{regIndicators}{matrix with ones and zeros specifying which parameters in regOn should be regularized. Must be of same size as the regularized matrix. 1 = regularized, 0 = not regularized. Alternatively, labels for the regularized parameters can be used (e.g. drift_eta1_eta2)}

\item{penalty}{type. Currently supported are lasso and ridge for optimization = approx and lasso for optimization = exact}

\item{elastic_alpha}{placehoder for elastic net. NOT YET IMPLEMENTED}

\item{elastic_gamma}{placehoder for elastic net. NOT YET IMPLEMENTED}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{returnFitIndices}{Boolean: should fit indices be returned?}

\item{cvSample}{cross-validation sample. Has to be of type mxData}

\item{autoCV}{Boolean: Should automatic cross-validation be used?}

\item{k}{number of cross-validation folds if autoCV = TRUE (k-fold cross-validation)}

\item{epsilon}{epsilon is used to transform the non-differentiable lasso penalty to a differentiable one if optimization = approx}

\item{zeroThresh}{threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx}

\item{cores}{how many computer cores should be used?}

\item{verbose}{0 (default), 1 for convergence plot, 2 for parameter convergence plot and line search progress}

\item{silent}{silent execution}

\item{progressBar}{Boolean: Should a progress bar be displayed}

\item{parallelProgressBar}{list: used internally to display progress when executing in parallel}

\item{regValues}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}
}
\description{
sets up an approximate regularized ctsem
}
\author{
Jannik Orzek
}
