% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_RAMM2LLCpptsem.R
\name{approx_RAMM2LLCpptsem}
\alias{approx_RAMM2LLCpptsem}
\title{approx_KalmanM2LLCpptsem}
\usage{
approx_RAMM2LLCpptsem(parameters, cpptsemmodel, failureReturns)
}
\arguments{
\item{parameters}{paramneter values}

\item{cpptsemmodel}{model from cpptsem}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}
}
\description{
computes gradients for an approximate optimization of regularized ctsem based on cpptsem with Full Information Maximum Likelihood objective
}
\author{
Jannik Orzek
}
