% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_RAMRegM2LLCpptsem.R
\name{approx_RAMRegM2LLCpptsem}
\alias{approx_RAMRegM2LLCpptsem}
\title{approx_RAMRegM2LLCpptsem}
\usage{
approx_RAMRegM2LLCpptsem(
  parameters,
  cpptsemmodel,
  adaptiveLassoWeights,
  N,
  lambda,
  regIndicators,
  epsilon,
  objective,
  failureReturns
)
}
\arguments{
\item{parameters}{parameter values}

\item{cpptsemmodel}{model returned from cpptsem}

\item{adaptiveLassoWeights}{vector with weights of the adaptive lasso}

\item{N}{sample size}

\item{lambda}{tuning parameter lambda}

\item{regIndicators}{string vector with names of regularized parameters}

\item{epsilon}{tuning parameter for epsL1 approximation}

\item{objective}{ML or Kalman}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}
}
\description{
approximates the regularized likelihood function using cpptsem and Full Information Maximum Likelihood
}
\author{
Jannik Orzek
}
