% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{armaExpmat}
\alias{armaExpmat}
\title{armaExpmat}
\usage{
armaExpmat(m)
}
\arguments{
\item{m}{symmetric matrix for which the matrix exponential should be computed}
}
\value{
matrix
}
\description{
Computes the matrix exponential using RcppArmadillo
}
