% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_helper.R
\name{checkAutoKFold}
\alias{checkAutoKFold}
\title{checkAutoKFold}
\usage{
checkAutoKFold(ctInit, regCtsemObject, threshold, testIC)
}
\arguments{
\item{ctInit}{init object from ctsem}

\item{regCtsemObject}{object from regCtsem}
}
\description{
used in testthat to check the automatic k-fold cross-validation feature of regCtsem
}
