% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIdentification}
\alias{checkIdentification}
\title{checkIdentification}
\usage{
checkIdentification(model, pars)
}
\arguments{
\item{model}{model of type ctsemFit, Rcpp_cpptsemRAMmodel, or MxRAMModel. WARNING: Currently only supported for RAM based models}

\item{pars}{named vector with parameter values}
}
\value{
TRUE if identified, FALSE otherwise
}
\description{
Checks the local identification of a model around the provided parameters
}
\details{
The function is based on the procedure described in Huang, P.-H. (2020). lslx: Semi-Confirmatory Structural Equation Modeling via Penalized Likelihood. Journal of Statistical Software, 93(7). https://doi.org/10.18637/jss.v093.i07
}
