% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkNonConvexity}
\alias{checkNonConvexity}
\title{checkNonConvexity}
\usage{
checkNonConvexity(
  model,
  lowerBound = NULL,
  upperBound = NULL,
  nSample = 50,
  sampleWith = "runif"
)
}
\arguments{
\item{model}{model of type ctsemFit, Rcpp_cpptsemRAMmodel, Rcpp_cpptsemKalmanModel, or MxRAMModel}

\item{lowerBound}{vector of the same length as the number of parameters in the model specifying the lower bound for each parameter}

\item{upperBound}{vector of the same length as the number of parameters in the model specifying the upper bound for each parameter}

\item{nSample}{number of sample points between the lower and upper bound}

\item{sampleWith}{possible are runif (random points between lower and upper bound) and seq (equidistant points between lower and upper bound)}
}
\value{
returns the generated samplePoints, fitAtSamplePoints which gives the -2log-Lokelihood at the sample points, a logical vector indicating if the midpoint non-convexity was given for the comparison of these sample points (isNonConvex) and the difference f(x1) + f(x2) - 2*f((x1 + x2)/2) for each combination of sample points. This difference should be positive for all comparisons!


}
\description{
checks the non-convexity of a provided model within the provided parameter bounds numerically.
}
\details{
This is a very simple implementation of the concept described in Tamura, K., & Gallagher, M. (2019). Quantitative measure of nonconvexity for black-box continuous functions. Information Sciences, 476, 64–82. https://doi.org/10.1016/j.ins.2018.10.009
The basic idea is as follows: The function generates nSample points within the provided parameter bounds. For each combination of these sample points the following condition is evaluated:
f(x1) + f(x2) >= 2*f((x1 + x2)/2)  (midpoint convexity, see Equation 2 in Tamura, K., & Gallagher, M., 2019)
This condition should evaluate to TRUE for all possible combinations of x1 and x2. If it evaluates to FALSE, this implies that the model fitting function is not convex.
WARNING: Even if all evaluations indicate that midpoint convexity is given, this is no proof that the function is convex! However, if a counter-example is found (i.e., at least one evaluation results in FALSE), this implies
that the function might be nonconvex. Keep in mind that this is only a numerical check and susceptible to numerical errors
}
