% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkNonConvexity3D}
\alias{checkNonConvexity3D}
\title{checkNonConvexity3D}
\usage{
checkNonConvexity3D(
  model,
  parnames,
  lowerBound1,
  upperBound1,
  lowerBound2,
  upperBound2,
  nSamples
)
}
\arguments{
\item{model}{model of type ctsemFit, Rcpp_cpptsemRAMmodel, Rcpp_cpptsemKalmanModel, or MxRAMModel}

\item{parnames}{vector of length 2 with the names of the parameters for which the likelihood surface should be plotted}

\item{lowerBound1}{double: lower bound for the first parameter}

\item{upperBound1}{double: upper bound for the first parameter}

\item{lowerBound2}{double: lower bound for the second parameter}

\item{upperBound2}{double: upper bound for the second parameter}

\item{nSample}{number of sample points between the lower and upper bound}
}
\value{
returns the a list with arguments to pass to plot3D::persp (e.g. do.call(args, plot3D::persp))


}
\description{
provides a 3D plot of the likelihood surface for two selected variables
}
\details{
The function generates nSample points within the provided parameter bounds. For each combination of sample point
}
