% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{checkSetup}
\alias{checkSetup}
\title{checkSetup}
\usage{
checkSetup(argsIn)
}
\arguments{
\item{argsIn}{list with arguments}
}
\description{
internal checks
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
