% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDiscreteCINTs}
\alias{computeDiscreteCINTs}
\title{computeDiscreteCINTs}
\usage{
computeDiscreteCINTs(
  discreteCINTUnique,
  DRIFTInverseValues,
  discreteDRIFTUnique,
  CINTValues
)
}
\arguments{
\item{discreteCINTUnique}{list with unique discrete continuous-time-intercepts}

\item{DRIFTInverseValues}{matrix with inverse of drift values}

\item{discreteDRIFTUnique}{list with unique discrete drifts}

\item{CINTValues}{continuous time intercepts}
}
\value{
list with updated discrete continuous-time-intercepts
}
\description{
Computes the discrete time latent intercept values (list with discreteCINT-names, dTs and results) given the inverse of the drift (DRIFTInverseValues), the discreteDRIFTUnique (list with discreteDRIFT-names, dTs and results) and CINTValues
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
