% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDiscreteDRIFTs}
\alias{computeDiscreteDRIFTs}
\title{computeDiscreteDRIFTs}
\usage{
computeDiscreteDRIFTs(DRIFTValues, discreteDRIFTUnique)
}
\arguments{
\item{DRIFTValues}{matrix with drift values}

\item{discreteDRIFTUnique}{list with discrete drift names, dTs and results}
}
\value{
list
}
\description{
Computes the discrete time drift matrices given the drift and discreteDRIFTUnique (list with discrete drift names, dTs and results)
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
