% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDiscreteTRAITs}
\alias{computeDiscreteTRAITs}
\title{computeDiscreteTRAITs}
\usage{
computeDiscreteTRAITs(discreteDRIFTUnique, discreteTRAITUnique)
}
\arguments{
\item{discreteDRIFTUnique}{list with discrete drift names, dTs and results}

\item{discreteTRAITUnique}{list with discrete trait names, dTs and results}
}
\value{
list with discrete traits
}
\description{
Computes the discrete time trait matrices given the discreteDRIFTUnique (list with discrete drift names, dTs and results) and discreteTRAITUnique (list with discrete trait names, dTs and results)
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
