% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDRIFTHASH}
\alias{computeDRIFTHASH}
\title{computeDRIFTHASH}
\usage{
computeDRIFTHASH(DRIFTValues)
}
\arguments{
\item{DRIFTValues}{matrix with drift values}
}
\value{
matrix with drift hash
}
\description{
Computes the computeDRIFTHASH given the drift. DRIFTHASH = DRIFT otimes I + I otimes DRIFT.
}
\details{
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
