% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeGroupM2LL}
\alias{computeGroupM2LL}
\title{computeGroupM2LL
Computes the -2log likelihood in a RAM model for a  group of people with identical missing structure given the sample size, number of non-missing variables (nObservedVariables),
observedMeans within this sample, observedCov within this sample, filtered expectedMeans, and filtered expectedCovariance.
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05}
\usage{
computeGroupM2LL(
  sampleSize,
  nObservedVariables,
  observedMeans,
  observedCov,
  expectedMeans,
  expectedCovariance
)
}
\arguments{
\item{sampleSize}{number of persons}

\item{nObservedVariables}{number of observed variables}

\item{observedMeans}{means of observed variables}

\item{observedCov}{observed covariance}

\item{expectedMeans}{expected means}

\item{expectedCovariance}{expected covariance}
}
\value{
double
}
\description{
computeGroupM2LL
Computes the -2log likelihood in a RAM model for a  group of people with identical missing structure given the sample size, number of non-missing variables (nObservedVariables),
observedMeans within this sample, observedCov within this sample, filtered expectedMeans, and filtered expectedCovariance.
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
