% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeIndividualM2LL}
\alias{computeIndividualM2LL}
\title{computeIndividualM2LL}
\usage{
computeIndividualM2LL(
  nObservedVariables,
  rawData,
  expectedMeans,
  expectedCovariance
)
}
\arguments{
\item{nObservedVariables}{number of non-missing variables}

\item{rawData}{x}

\item{expectedMeans}{filtered expectedMeans}

\item{expectedCovariance}{filtered expectedCovariance}
}
\description{
Computes the -2log likelihood in a RAM model for a  single person given the number of non-missing variables (nObservedVariables), x (rawData), filtered expectedMeans, and filtered expectedCovariance.
}
\keyword{internal}
