% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeKalmanLatentCovariancePrediction}
\alias{computeKalmanLatentCovariancePrediction}
\title{computeKalmanLatentCovariancePrediction}
\usage{
computeKalmanLatentCovariancePrediction(
  discreteDRIFTValues,
  previousCovariances,
  discreteDIFFUSIONValues
)
}
\arguments{
\item{discreteDRIFTValues}{discrete drift values}

\item{previousCovariances}{covariances of previous time point}

\item{discreteDIFFUSIONValues}{discrete diffusion values}
}
\value{
predicted covariance
}
\description{
Computes the prediction step for the residual covariance matrix of the latent states when using the Kalman filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
