% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeKalmanLatentStatePrediction}
\alias{computeKalmanLatentStatePrediction}
\title{computeKalmanLatentStatePrediction
Computes the prediction step for the latent states when using the Kalman filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05}
\usage{
computeKalmanLatentStatePrediction(
  discreteDRIFTValues,
  previousStates,
  discreteCINTValues
)
}
\arguments{
\item{discreteDRIFTValues}{discrete drift values}

\item{previousStates}{states of previous time point}

\item{discreteCINTValues}{discrete continuous-time-intercept values}
}
\value{
vector with predicted states
}
\description{
computeKalmanLatentStatePrediction
Computes the prediction step for the latent states when using the Kalman filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
