% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeKalmanManifestCovariancePrediction}
\alias{computeKalmanManifestCovariancePrediction}
\title{computeKalmanManifestCovariancePrediction}
\usage{
computeKalmanManifestCovariancePrediction(
  LAMBDA,
  predictedLatentCovariances,
  currentMANIFESTVAR
)
}
\arguments{
\item{LAMBDA}{loadings matrix}

\item{predictedLatentCovariances}{predicted latent covariances}

\item{currentMANIFESTVAR}{current manifest covariance}
}
\description{
Computes the predicted manifest residual covariance matrix given the latent residual covariance when using the Kalman filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
