% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeKalmanManifestPrediction}
\alias{computeKalmanManifestPrediction}
\title{computeKalmanManifestPrediction}
\usage{
computeKalmanManifestPrediction(
  LAMBDAValues,
  predictedStates,
  MANIFESTMEANSValues
)
}
\arguments{
\item{LAMBDAValues}{matrix with loadings}

\item{predictedStates}{vector with predicted latent states}

\item{MANIFESTMEANSValues}{vector with means of manifest variables}
}
\value{
predicted manifest values
}
\description{
Computes the Prediction for the manifest variables given the predicted latent states when using the Kalman filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
