% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeRAMExpectedCovariance}
\alias{computeRAMExpectedCovariance}
\title{computeRAMExpectedCovariance}
\usage{
computeRAMExpectedCovariance(F, A, S)
}
\arguments{
\item{F}{filter matrix}

\item{A}{direct effects matrix}

\item{S}{undirected effects matrix}
}
\value{
matrix with model implied covariance
}
\description{
Computes the expected covariance in a RAM model give filter matrix F, directed effects matrix A, and covariance matrix S.
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
