% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeRAMExpectedMeans}
\alias{computeRAMExpectedMeans}
\title{computeRAMExpectedMeans}
\usage{
computeRAMExpectedMeans(F, A, M)
}
\arguments{
\item{F}{filter matrix F}

\item{A}{matrix with directed effects}

\item{M}{vector with means}
}
\value{
matrix with implied means
}
\description{
Computes the expected means in a RAM model give filter matrix F, directed effects matrix A, and means vector M.
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
