% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeRAMM2LL}
\alias{computeRAMM2LL}
\title{computeRAMM2LL}
\usage{
computeRAMM2LL(RAMdata, expectedMeans, expectedCovariance)
}
\arguments{
\item{RAMdata}{list with data set}

\item{expectedMeans}{vector with model implied means}

\item{expectedCovariance}{matrix with model implied covariance}
}
\value{
double
}
\description{
Computes the -2 log Likelihood
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
