% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{computeStandardErrors}
\alias{computeStandardErrors}
\title{computeStandardErrors}
\usage{
computeStandardErrors(cpptsemObject, objective, sampleSize)
}
\arguments{
\item{cpptsemObject}{fitted cpptsemObject}

\item{objective}{Kalman or ML}

\item{sampleSize}{sample size}
}
\description{
computes standard errors for a cpptsem object
}
\author{
Jannik H. Orzek
}
