% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{computeStandardErrorsDelta}
\alias{computeStandardErrorsDelta}
\title{computeStandardErrorsDelta}
\usage{
computeStandardErrorsDelta(
  cpptsemObject,
  objective,
  eps = (1.1 * 10^(-16))^(1/3)
)
}
\arguments{
\item{cpptsemObject}{fitted cpptsemObject}

\item{objective}{Kalman or ML}

\item{eps}{epsilon for numerical approximation of derivatives}
}
\description{
computes standard errors for the transformed parameters of a cpptsem object
}
\author{
Jannik H. Orzek
}
\keyword{internal}
