% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{computeStandardErrorsRaw}
\alias{computeStandardErrorsRaw}
\title{computeStandardErrorsRaw}
\usage{
computeStandardErrorsRaw(cpptsemObject, objective)
}
\arguments{
\item{cpptsemObject}{fitted cpptsemObject}

\item{objective}{Kalman or ML}
}
\description{
computes standard errors of the raw parameters for a cpptsem object
}
\author{
Jannik H. Orzek
}
