% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{constructDatasetKalman}
\alias{constructDatasetKalman}
\title{constructDatasetKalman}
\usage{
constructDatasetKalman(wideData)
}
\arguments{
\item{wideData}{dataset in wide format compatible with ctsem}
}
\description{
Separates data and time intervals. Computes the number of unique time intervals and the unqiue missingness patterns
}
