% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlDocumentation.R
\name{controlOptimx}
\alias{controlOptimx}
\title{controlOptimx}
\usage{
controlOptimx(
  package = "optimx",
  failureReturns = .Machine$double.xmax/2,
  hess = NULL,
  lower = -Inf,
  upper = Inf,
  method = "L-BFGS-B",
  hessian = FALSE,
  itnmax = 200,
  control = list(dowarn = FALSE, kkt = TRUE, maxit = 200)
)
}
\arguments{
\item{package}{set to "optimx"}

\item{failureReturns}{what will the fitting function return if the current points are impossible? Depends on the method used}

\item{hess}{Should the Hessian be computed at the solution}

\item{lower}{lower bounds for paramters}

\item{upper}{upper bounds for parameters}

\item{method}{see ?optimx}

\item{itnmax}{maximal number of iterations}

\item{control}{control passed to optimx}
}
\description{
list with settings used for optimization with optimx
}
