% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlDocumentation.R
\name{controlRsolnp}
\alias{controlRsolnp}
\title{controlRsolnp}
\usage{
controlRsolnp(
  package = "Rsolnp",
  failureReturns = .Machine$double.xmax/2,
  eqfun = NULL,
  eqB = NULL,
  ineqfun = NULL,
  ineqLB = NULL,
  ineqUB = NULL,
  LB = NULL,
  UB = NULL,
  control = list(outer.iter = 50, trace = 0)
)
}
\arguments{
\item{package}{set to "Rsolnp"}

\item{failureReturns}{what will the fitting function return if the current points are impossible?}

\item{eqfun}{Equality constraints function. See ?Rsolnp::solnp}

\item{eqB}{Equality constraints. See ?Rsolnp::solnp}

\item{ineqfun}{Inequality constraints function. See ?Rsolnp::solnp}

\item{ineqLB}{Inequality constraints lower bound. See ?Rsolnp::solnp}

\item{ineqUB}{Inequality constraints upper bound. See ?Rsolnp::solnp}

\item{LB}{Lower bound. See ?Rsolnp::solnp}

\item{UB}{Upper bound. See ?Rsolnp::solnp}

\item{control}{control passed to Rsolnp}
}
\description{
list with settings used for optimization with Rsolnp
}
