% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCVFitTable.R
\name{createCVFitTable}
\alias{createCVFitTable}
\title{createCVFitTable}
\usage{
createCVFitTable(k, regValues)
}
\arguments{
\item{k}{number of cross-validation folds if autoCV = TRUE (k-fold cross-validation)}

\item{regValues}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}
}
\description{
sets up a table for CV fit values
}
\author{
Jannik Orzek
}
