% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{createKalmanMultiSubjectModel}
\alias{createKalmanMultiSubjectModel}
\title{createKalmanMultiSubjectModel}
\usage{
createKalmanMultiSubjectModel(
  ctsemObject,
  dataset,
  useOptimizer,
  silent = FALSE,
  KalmanStartValues = NULL
)
}
\arguments{
\item{ctsemObject}{object of type ctsemInit from ctModel}

\item{dataset}{data set in wide format compatible to ctsem}

\item{useOptimizer}{Boolean: should the model be optimized}
}
\description{
Creates an mxModel Object for N >= 1 individuals using the Kalman filter
}
\details{
NOTE: Function located in file regCtsem.R
}
