% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyDataFrame}
\alias{deepCopyDataFrame}
\title{deepCopyDataFrame}
\usage{
deepCopyDataFrame(x)
}
\arguments{
\item{x}{Rcpp DataFrame}
}
\value{
copy of Rcpp DataFrame
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
