% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyList}
\alias{deepCopyList}
\title{deepCopyList}
\usage{
deepCopyList(x)
}
\arguments{
\item{x}{Rcpp List}
}
\value{
copy of Rcpp List
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
