% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyNumericMatrix}
\alias{deepCopyNumericMatrix}
\title{deepCopyNumericMatrix}
\usage{
deepCopyNumericMatrix(x)
}
\arguments{
\item{x}{Rcpp NumericMatrix}
}
\value{
copy of Rcpp NumericMatrix
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
