% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyNumericVector}
\alias{deepCopyNumericVector}
\title{deepCopyNumericVector}
\usage{
deepCopyNumericVector(x)
}
\arguments{
\item{x}{Rcpp NumericVector}
}
\value{
copy of Rcpp NumericVector
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
