% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyStringMatrix}
\alias{deepCopyStringMatrix}
\title{deepCopyStringMatrix}
\usage{
deepCopyStringMatrix(x)
}
\arguments{
\item{x}{Rcpp StringMatrix}
}
\value{
copy of Rcpp StringMatrix
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
