% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deepCopyStringVector}
\alias{deepCopyStringVector}
\title{deepCopyStringVector}
\usage{
deepCopyStringVector(x)
}
\arguments{
\item{x}{Rcpp StringVector}
}
\value{
copy of Rcpp StringVector
}
\description{
deep copy of Rcpp data
}
\keyword{internal}
