% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_armijoLineSearch}
\alias{exact_armijoLineSearch}
\title{exact_armijoLineSearch}
\usage{
exact_armijoLineSearch(
  gradientModel,
  adaptiveLassoWeights,
  parameterLabels,
  regIndicators,
  lambda,
  newParameters,
  m2LLNew,
  newGradient,
  direction,
  c1,
  stepSize,
  differenceApprox,
  maxIter_line
)
}
\arguments{
\item{gradientModel}{mxObject for computing the derivarive of the likelihood with respect to the parameter estimates}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{parameterLabels}{names of the parameter estimates}

\item{regIndicators}{vector with names of parameters to regularize}

\item{lambda}{penalty value}

\item{newParameters}{parameter values of iteration k plus 1}

\item{m2LLNew}{-2 log likelihood of iteration k plus 1}

\item{newGradient}{gradients of iteration k plus 1}

\item{direction}{vector with updates to parameter estimates}

\item{c1}{tuning parameter for Armijo condition}

\item{stepSize}{Initial stepsize of the outer iteration (theta_{k+1} = oldParameters + Stepsize \* Stepdirection)}

\item{differenceApprox}{which approximation for the gradients should be used? Recommended is central}

\item{maxIter_line}{maximal number of iterations for line search}
}
\description{
performs Armijo line search
}
\details{
NOTE: Function located in file GLMNET.R
}
\keyword{internal}
