% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_getBFGS}
\alias{exact_getBFGS}
\title{exact_getBFGS}
\usage{
exact_getBFGS(
  oldParameters,
  oldGradients,
  oldHessian,
  newParameters,
  newGradient,
  cautious = TRUE,
  hessianEps = 0.001
)
}
\arguments{
\item{oldParameters}{Theta at iteration k}

\item{oldGradients}{Gradients of the likelihood function at iteration k}

\item{oldHessian}{Hessian of the likelihood function at iteration k}

\item{newParameters}{Theta at iteration k+1}

\item{newGradient}{Gradients of the likelihood function at iteration k+1}

\item{cautious}{boolean: should the update be skipped if it would result in a non positive definite Hessian?}

\item{hessianEps}{controls when the update of the Hessian approximation is skipped}
}
\description{
computes the BFGS Hessian approximation
}
\details{
NOTE: Function located in file GLMNET.R
}
