% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_getCppGradients.R
\name{exact_getCppGradients}
\alias{exact_getCppGradients}
\title{exact_getCppGradients}
\usage{
exact_getCppGradients(cppmodel, objective)
}
\arguments{
\item{cppmodel}{model of type cpptsem}

\item{objective}{ml or Kalman}
}
\description{
exact_getCppGradients will try to compute gradients with decreasing precision starting from the default in OpenMx. Sometimes the gradients will result in NA for a very specific setting of the precisions. Then it can help to slightly alter the precision
}
