% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_Optimization.R
\name{exact_getCVFit}
\alias{exact_getCVFit}
\title{exact_getCVFit}
\usage{
exact_getCVFit(
  objective,
  ctsemObject,
  mxObject,
  parameterLabels,
  parameterValuesTable,
  regValues,
  cvSample
)
}
\arguments{
\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{ctsemObject}{if objective = "ML": Fitted object of class ctsem. If you want to use objective = "Kalman", pass an object of type ctsemInit from ctModel}

\item{mxObject}{Fitted object of class MxObject}

\item{parameterLabels}{labels of optimized parameters}

\item{parameterValuesTable}{table with parameter values+}

\item{regValues}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}

\item{cvSample}{cross-validation sample. Has to be of type mxData}
}
\description{
computes cross-validation fit for exact optimization
}
\details{
NOTE: Function located in file exact_Optimization.R
}
\author{
Jannik Orzek
}
