% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getFitIndicesWithTarget}
\alias{exact_getFitIndicesWithTarget}
\title{exact_getFitIndicesWithTarget}
\usage{
exact_getFitIndicesWithTarget(
  parameterLabels,
  regIndicators,
  fitAndParameters,
  targetVector,
  lambdas,
  sampleSize
)
}
\arguments{
\item{parameterLabels}{labels of optimized parameters}

\item{regIndicators}{Vector with names of regularized parameters}

\item{fitAndParameters}{table with fit and parameter values}

\item{targetVector}{named vector with values towards which the parameters are regularized}

\item{lambdas}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}

\item{sampleSize}{sample size}

\item{mxObject}{Fitted object of class MxObject}
}
\description{
computes fit indices for optimization = "exact"
}
\details{
NOTE: Function located in file exact_optimization.R
}
\author{
Jannik Orzek
}
