% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getFlatStdizer}
\alias{exact_getFlatStdizer}
\title{exact_getFlatStdizer}
\usage{
exact_getFlatStdizer(T0VAR, thetaNames)
}
\arguments{
\item{T0VAR}{matrix with T0VAR values}

\item{thetaNames}{vector with parameter names}
}
\description{
returns the standardizer for the standardized drift parameters if standardizeDrift = TRUE. Computes T0SD_predictor/T0SD_dependent
}
\details{
NOTE: Function located in file exact_optimization.R
}
