% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getLambda}
\alias{exact_getLambda}
\title{exact_getLambda}
\usage{
exact_getLambda(lambda, theta, regIndicators, adaptiveLassoWeights = NULL)
}
\arguments{
\item{lambda}{Penaltiy value}

\item{regIndicators}{Names of regularized parameters}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}
}
\description{
computes sum(lambda*abs(regularized Values))
}
\details{
NOTE: Function located in file exact_optimization.R
}
