% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getPenaltyValue}
\alias{exact_getPenaltyValue}
\title{exact_getPenaltyValue}
\usage{
exact_getPenaltyValue(lambda, theta, regIndicators, adaptiveLassoWeights)
}
\arguments{
\item{lambda}{Penalty value}

\item{theta}{parameter values}

\item{regIndicators}{Names of regularized parameters}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}
}
\description{
computes sum(lambda*abs(regularized Values))
}
\details{
NOTE: Function located in file exact_optimization.R
}
