% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getPenaltyValueWithTarget}
\alias{exact_getPenaltyValueWithTarget}
\title{exact_getPenaltyValueWithTarget}
\usage{
exact_getPenaltyValueWithTarget(
  lambda,
  theta,
  regIndicators,
  targetVector,
  adaptiveLassoWeights = NULL
)
}
\arguments{
\item{lambda}{Penalty value}

\item{regIndicators}{Names of regularized parameters}

\item{targetVector}{named vector with values towards which the parameters are regularized}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}
}
\description{
computes sum(lambda*abs(regularized Values))
}
\details{
NOTE: Function located in file exact_optimization.R
}
