% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_Optimization.R
\name{exact_getRegValue}
\alias{exact_getRegValue}
\title{exact_getRegValue}
\usage{
exact_getRegValue(lambda, theta, regIndicators, adaptiveLassoWeights = NULL)
}
\arguments{
\item{lambda}{Penaltiy value}

\item{regIndicators}{Names of regularized parameters}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}
}
\description{
computes sum(lambda*abs(regularized Values))
}
\details{
NOTE: Function located in file exact_Optimization.R
}
