% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_Optimization.R
\name{exact_getT0VAR}
\alias{exact_getT0VAR}
\title{exact_getT0VAR}
\usage{
exact_getT0VAR(mxObject, d)
}
\arguments{
\item{mxObject}{mxObject with old parameter values}

\item{d}{vector with updates to parameter values}

\item{stationarity}{set to TRUE, if stationaritiy is assumed for the T0VAR}
}
\description{
computes the updated T0VAR given the old parameter values in an mxObject and the updates to the parameters in a vector d
}
\details{
NOTE: Function located in file exact_Optimization.R
}
