% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_GLMNETLineSearch}
\alias{exact_GLMNETLineSearch}
\title{exact_GLMNETLineSearch}
\usage{
exact_GLMNETLineSearch(
  cpptsemObject,
  objective,
  adaptiveLassoWeights,
  parameterLabels,
  regIndicators,
  lambda,
  newParameters,
  m2LLNew,
  newGradient,
  oldHessian,
  direction,
  stepSize,
  sig,
  gam,
  maxIter_line
)
}
\arguments{
\item{cpptsemObject}{cpptsem object which specifies how the gradients of the likelihood-function are computed}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{parameterLabels}{names of the parameter estimates}

\item{regIndicators}{vector with names of parameters to regularize}

\item{lambda}{penalty value}

\item{newParameters}{parameter values of iteration k plus 1}

\item{m2LLNew}{-2 log likelihood of iteration k plus 1}

\item{newGradient}{gradients of iteration k plus 1}

\item{oldHessian}{Hessian approximation}

\item{direction}{vector with updates to parameter estimates}

\item{stepSize}{Initial stepsize of the outer iteration (theta_{k+1} = oldParameters + Stepsize \* Stepdirection)}

\item{sig}{only relevant when lineSearch = 'GLMNET'. Controls the sigma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421.}

\item{gam}{Controls the gamma parameter in Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421. Defaults to 0.}

\item{maxIter_line}{maximal number of iterations for line search}
}
\description{
performs the line search procedure described by Yuan, G.-X., Ho, C.-H., & Lin, C.-J. (2012). An improved GLMNET for l1-regularized logistic regression. The Journal of Machine Learning Research, 13, 1999–2030. https://doi.org/10.1145/2020408.2020421 Equation 20.
}
\details{
NOTE: Function located in file GLMNET.R
}
\keyword{internal}
