% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_Optimization.R
\name{exact_initialHessian}
\alias{exact_initialHessian}
\title{exact_initialHessian}
\usage{
exact_initialHessian(mxObject, approximationType, estimatedHessian)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject}

\item{approximationType}{which Hessian should be used? Currently available are "ident" for an identity matrix and "OpenMx" for the Hessian from OpenMx Gradient Descent (recommended)}

\item{estimatedHessian}{estimated Hessian from OpenMx}
}
\description{
computes an initial positive-definite Hessian matrix
}
\details{
NOTE: Function located in file exact_Optimization.R
}
\author{
Jannik Orzek
}
