% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_innerGLMNET}
\alias{exact_innerGLMNET}
\title{exact_innerGLMNET}
\usage{
exact_innerGLMNET(
  adaptiveLassoWeights,
  parameterLabels,
  regIndicators,
  lambda,
  newParameters,
  newGradient,
  HessianNew,
  maxIter_in,
  eps_in
)
}
\arguments{
\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{parameterLabels}{Vector with names of theta-parameters}

\item{regIndicators}{Names of regularized parameters}

\item{lambda}{Penalty value}

\item{newParameters}{Theta at iteration k+1}

\item{newGradient}{Gradients of the likelihood function at iteration k+1}

\item{HessianNew}{Hessian of the likelihood function at iteration k+1}

\item{maxIter_in}{Maximal number of iterations of the inner optimization algorithm}

\item{eps_in}{Stopping criterion for the inner iterations}
}
\description{
performs the inner optimization routine of GLMNET. exact_innerGLMNET returns a direction vector for the next step in the outer optimization routine.
}
\details{
NOTE: Function located in file GLMNET.R
}
\keyword{internal}
