% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_weakWolfeLineSearch}
\alias{exact_weakWolfeLineSearch}
\title{exact_weakWolfeLineSearch}
\usage{
exact_weakWolfeLineSearch(
  gradientModel,
  objective,
  gradientModelcpp = NULL,
  adaptiveLassoWeights,
  thetaNames,
  regIndicators,
  lambda,
  theta_kp1,
  m2LL_kp1,
  g_kp1,
  d,
  c1 = 1e-04,
  c2 = 0.9,
  stepSize = 1,
  differenceApprox = "central",
  maxIter_line = 100,
  verbose = 0,
  eps_WW = 1e-04,
  eps_numericDerivative = (1.1 * 10^(-16))^(1/3)
)
}
\arguments{
\item{gradientModel}{mxObject for computing the derivarive of the likelihood with respect to the parameter estimates}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{thetaNames}{names of the parameter estimates}

\item{regIndicators}{vector with names of parameters to regularize}

\item{lambda}{penalty value}

\item{theta_kp1}{parameter values of iteration k plus 1}

\item{m2LL_kp1}{-2 log likelihood of iteration k plus 1}

\item{g_kp1}{gradients of iteration k plus 1}

\item{d}{vector with updates to parameter estimates}

\item{c1}{tuning parameter for Armijo condition}

\item{c2}{c2 constant for lineSearch. This constant controls the Curvature condition}

\item{stepSize}{Initial stepsize of the outer iteration (theta_{k+1} = theta_k + Stepsize \* Stepdirection)}

\item{differenceApprox}{which approximation for the gradients should be used? Recommended is central}

\item{maxIter_line}{maximal number of iterations for line search}

\item{verbose}{0 (default), 1 for convergence plot, 2 for parameter convergence plot and line search progress}

\item{eps_numericDerivative}{controls the precision of the central gradient approximation. The default (1.1 * 10^(-16))^(1/3) is derived in Nocedal, J., & Wright, S. J. (2006). Numerical optimization (2nd ed), p. 197}

\item{epsWW}{additional parameter to shorten the search if the upper and lower bound are already very close to each other}
}
\description{
NOTE: Function located in file GLMNET.R
}
