% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{extractCtsemMatrices}
\alias{extractCtsemMatrices}
\title{extractCtsemMatrices}
\usage{
extractCtsemMatrices(mxObject, nlatent, nmanifest, silent = FALSE)
}
\arguments{
\item{mxObject}{mxObject from ctsemOMX}

\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}
}
\description{
Extract DRIFT, DIFFUSION, etc from an mxObject
}
