% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{extractOptimx}
\alias{extractOptimx}
\title{extractOptimx}
\usage{
extractOptimx(parameterLabels, opt)
}
\arguments{
\item{parameterLabels}{vector with parameter labels}

\item{opt}{result from calling psydiffOptimx}
}
\description{
sets the model parameters to the best values obtained from optimx
}
