% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{extractParameterTableFromMx}
\alias{extractParameterTableFromMx}
\title{extractParameterTableFromMx}
\usage{
extractParameterTableFromMx(mxObject)
}
\arguments{
\item{mxObject}{Object of type mxObject}
}
\description{
wrapper for the omxLocateParameters function from OpenMx
returns a data frame with parameter labels, location and values
}
\keyword{internal}
