% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillA}
\alias{fillA}
\title{fillA}
\usage{
fillA(
  A,
  hasDiscreteDRIFTUnique,
  discreteDRIFTUnique,
  hasDiscreteTRAITUnique,
  discreteTRAITUnique,
  LAMBDA,
  AParameterIndicators
)
}
\arguments{
\item{A}{matrix with directed effects}

\item{hasDiscreteDRIFTUnique}{boolean}

\item{discreteDRIFTUnique}{list with discrete drift names, dTs and results}

\item{hasDiscreteTRAITUnique}{boolean}

\item{discreteTRAITUnique}{list with discrete trait names, dTs and results}

\item{LAMBDA}{loadings}

\item{AParameterIndicators}{tells fillA where to put the discreteDRIFT, discreteTRAIT, and LAMBDA}
}
\description{
Fills the A matrix (directed effects) given the A matrix, the discreteDRIFTUnique (list with discrete drift names, dTs and results),
discreteTRAITUnique (list with discrete trait names, dTs and results), LAMBDA (loadings), and AParameterIndicators (tells fillA where to put the discreteDRIFT, discreteTRAIT, and LAMBDA)
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
