% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillM}
\alias{fillM}
\title{fillM}
\usage{
fillM(
  M,
  T0MEANS,
  MANIFESTMEANS,
  hasDiscreteCINTUnique,
  discreteCINTUnique,
  cppMParameterIndicators
)
}
\arguments{
\item{M}{vector with means}

\item{T0MEANS}{vector with initial means}

\item{MANIFESTMEANS}{manifest means}

\item{hasDiscreteCINTUnique}{boolean}

\item{discreteCINTUnique}{list with discrete continuous time intercepts, dTs and results}

\item{cppMParameterIndicators}{tells fillM where to put the MANIFESTMEANS and discreteCINT}
}
\description{
Fills the M vector (intercepts) given the M vector, the MANIFESTMEANS,
discreteCINTUnique (list with discrete continuous time intercepts, dTs and results), and cppMParameterIndicators (tells fillM where to put the MANIFESTMEANS and discreteCINT)
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
