% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fillS}
\alias{fillS}
\title{fillS}
\usage{
fillS(
  S,
  T0VAR,
  MANIFESTVAR,
  hasDiscreteTRAITUnique,
  TRAITVAR,
  hasDiscreteDIFFUSIONUnique,
  discreteDIFFUSIONUnique,
  SParameterIndicators
)
}
\arguments{
\item{S}{matrix with undirected effects}

\item{T0VAR}{matrix with initial covariances}

\item{MANIFESTVAR}{manifest variance}

\item{hasDiscreteTRAITUnique}{boolean}

\item{TRAITVAR}{trait variance}

\item{hasDiscreteDIFFUSIONUnique}{boolean}

\item{discreteDIFFUSIONUnique}{list with discrete diffusion names, dTs and results}

\item{SParameterIndicators}{tells fillS where to put the discreteDIFFUSIONUnique and MANIFESTVAR}
}
\description{
Fills the S matrix (covariances) given the S matrix, the discreteDIFFUSIONUnique (list with discrete diffusion names, dTs and results),
MANIFESTVAR (manifest variance), and cppSParameterIndicators (tells fillS where to put the discreteDIFFUSIONUnique and MANIFESTVAR)
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
