% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{fitCpptsem}
\alias{fitCpptsem}
\title{fitCpptsem}
\usage{
fitCpptsem(
  parameterValues,
  cpptsemObject,
  objective,
  free = labeledFree(parameterValues),
  failureReturns
)
}
\arguments{
\item{parameterValues}{vector with labeled parameter values}

\item{cpptsemObject}{cpptsem object}

\item{objective}{ML or Kalman}

\item{free}{vector of same length as cpptsemObject$getParameterValues, where for each element it is specified if the parameter is freely estimated (workaround for fixing parameters)}

\item{failureReturns}{e.g., NA, Inf, ... adapt to optimizer settings}
}
\description{
fits a cpptsem. Returns -2 log Likelihood
}
\keyword{internal}
