% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getAdaptiveLassoWeights}
\alias{getAdaptiveLassoWeights}
\title{getAdaptiveLassoWeights}
\usage{
getAdaptiveLassoWeights(
  cpptsemObject,
  penalty,
  adaptiveLassoWeights,
  standardizeDrift
)
}
\arguments{
\item{cpptsemObject}{Fitted object of class cpptsem}

\item{penalty}{type}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{standardizeDrift}{Boolean: Should Drift parameters be standardized automatically using T0VAR?}
}
\description{
Computes the adaptive lasso weights
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
