% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getCurvedLambda}
\alias{getCurvedLambda}
\title{getCurvedLambda}
\usage{
getCurvedLambda(maxLambda, lambdasAutoCurve, lambdasAutoLength)
}
\arguments{
\item{maxLambda}{maximal lambda value}

\item{lambdasAutoCurve}{controls the curve. A value close to 1 will result in a linear increase, larger values in lambdas more concentrated around 0}

\item{lambdasAutoLength}{number of lambda values to generate}
}
\description{
generates lambda values between 0 and lambdaMax using the function described here: https://math.stackexchange.com/questions/384613/exponential-function-with-values-between-0-and-1-for-x-values-between-0-and-1.
}
\examples{
plot(getCurvedLambda(10,1,100))
plot(getCurvedLambda(10,5,100))
plot(getCurvedLambda(10,100,100))
}
