% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getFinalModel}
\alias{getFinalModel}
\title{getFinalModel}
\usage{
getFinalModel(regCtsemObject, criterion = NULL)
}
\arguments{
\item{regCtsemObject}{fitted regularized continuous time model}

\item{criterion}{select a criterion. Possible are AIC, BIC, cvM2LL}
}
\description{
returns the final model for a regularized model
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
