% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getFinalParameters}
\alias{getFinalParameters}
\title{getFinalParameters}
\usage{
getFinalParameters(regCtsemObject, criterion = NULL, raw = FALSE)
}
\arguments{
\item{regCtsemObject}{fitted regularized continuous time model}

\item{criterion}{select a criterion. Possible are AIC, BIC, cvM2LL}

\item{raw}{boolean: should the raw parameters be returned (raw = TRUE) or should the log-transformed variances be re-transformed (raw = FALSE)}
}
\description{
returns the final parameters for a regularized model
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
