% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getFlatStdizer}
\alias{getFlatStdizer}
\title{getFlatStdizer}
\usage{
getFlatStdizer(VARIs, driftLabels)
}
\arguments{
\item{VARIs}{matrix with variances for standardization values}

\item{driftLabels}{vector with drift names}
}
\description{
returns the standardizer for the standardized drift parameters. Computes SD_predictor/SD_dependent
}
\details{
NOTE: Function located in file regCtsem.R
}
\keyword{internal}
