% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getMaxLambda}
\alias{getMaxLambda}
\title{getMaxLambda}
\usage{
getMaxLambda(
  cpptsemObject,
  objective,
  regIndicators,
  targetVector,
  adaptiveLassoWeights,
  nMultistart
)
}
\arguments{
\item{cpptsemObject}{Fitted object of class cpptsem}

\item{objective}{ML or Kalman}

\item{regIndicators}{Labels of the regularized parameters (e.g. drift_eta1_eta2)}

\item{targetVector}{vector with target values}

\item{adaptiveLassoWeights}{weights for the adaptive lasso. If auto, defaults to the unregularized parameter estimates.}

\item{nMultistart}{number of multi-start iterations}
}
\description{
computes an approximation of the lowest lambda which will set all regularized parameters to zero. This function is adapted from Murphy (2012) Machine learning: a probabilistic perspective. See p. 434 for more details.
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
\keyword{internal}
