% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getMaxRegValueCV}
\alias{getMaxRegValueCV}
\title{getMaxRegValueCV}
\usage{
getMaxRegValueCV(
  ctsemObject,
  mxObject,
  fullData,
  KalmanStartValues,
  regOn,
  regIndicators,
  penalty,
  adaptiveLassoWeights,
  standardizeDrift,
  k,
  cvFolds,
  trainSets,
  scaleLambdaWithN,
  objective,
  optimization,
  differenceApprox = "central"
)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{fullData}{Dataset for all samples combined}

\item{KalmanStartValues}{Optional starting values for the parameters when using Kalman filter}

\item{regOn}{string specifying which matrix should be regularized. Currently only supports DRIFT}

\item{regIndicators}{Labels of the regularized parameters (e.g. drift_eta1_eta2)}

\item{penalty}{type. Currently supported are lasso and ridge for optimization = approx and lasso and adaptiveLasso for optimization = exact}

\item{adaptiveLassoWeights}{weights for the adaptive lasso. If auto, defaults to the unregularized parameter estimates.}

\item{standardizeDrift}{Boolean: Should Drift parameters be standardized automatically using T0VAR?}

\item{k}{number of cross-validation folds if autoCV = TRUE (k-fold cross-validation)}

\item{cvFolds}{list of numeric vectors indicating which data row in fullData belongs to which cv sample}

\item{trainSets}{empty list of the same length as k. The data sets for training the models will be saved here}

\item{scaleLambdaWithN}{Boolean: Should the penalty value be scaled with the sample size? True is recommended, as the likelihood is also sample size dependent}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{optimization}{which optimization procedure should be used. Possible are  "exact" or "approx".}

\item{differenceApprox}{Which approximation should be used for calculating the gradients in the gradientModel. central is recommended}
}
\description{
computes an approximation of the lowest regValue which will set all regularized parameters to zero. This function is adapted from Murphy (2012) Machine learning: a probabilistic perspective. See p. 434 for more details.
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
