% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getParameterEstimates}
\alias{getParameterEstimates}
\title{getParameterEstimates}
\usage{
getParameterEstimates(regCtsemObject, parameterEstimatesRaw)
}
\arguments{
\item{regCtsemObject}{regCtsemObject}

\item{parameterEstimatesRaw}{raw parameter estimates from regularized ctsem}
}
\description{
computes the parameters given the raw parameter estimates
}
\details{
NOTE: Function located in file regCtsem.R
}
