% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{getRowsAndColsOfNonMissing}
\alias{getRowsAndColsOfNonMissing}
\title{getRowsAndColsOfNonMissing}
\usage{
getRowsAndColsOfNonMissing(uniqueMissingPattern)
}
\arguments{
\item{uniqueMissingPattern}{unique missingness patterns}
}
\description{
Build indicators for missingness patterns
}
