% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getT0VAR}
\alias{getT0VAR}
\title{getT0VAR}
\usage{
getT0VAR(mxObject)
}
\arguments{
\item{mxObject}{mxObject}
}
\description{
extracts the T0VAR from an mxObject
}
\details{
NOTE: Function located in file regCtsem.R
}
