% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getVarianceFromVarianceBase}
\alias{getVarianceFromVarianceBase}
\title{getVarianceFromVarianceBase}
\usage{
getVarianceFromVarianceBase(varianceBaseValues)
}
\arguments{
\item{varianceBaseValues}{matrix with base values}
}
\description{
Computes the covariance matrix given the covariancebase
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
