% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getVarianceFromVarianceBase2}
\alias{getVarianceFromVarianceBase2}
\title{getVarianceFromVarianceBase2}
\usage{
getVarianceFromVarianceBase2(varianceBaseValues)
}
\arguments{
\item{varianceBaseValues}{values of varianceBase}
}
\description{
returns variance from varianceBase
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
