% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getVariances}
\alias{getVariances}
\title{getVariances}
\usage{
getVariances(
  parameterEstimatesRaw,
  matName,
  baseMatName,
  parameterTable,
  nVariables,
  variableNames
)
}
\arguments{
\item{parameterEstimatesRaw}{raw parameter estimates from regularized ctsem}

\item{matName}{name of the new matrix}

\item{baseMatName}{name of the base matrix}

\item{parameterTable}{parameterTable}

\item{nVariables}{number of variables in the matrix}

\item{variableNames}{names of the variables}
}
\description{
computes the variances for varianceBase matrices
}
\details{
NOTE: Function located in file regCtsem.R
}
\keyword{internal}
