% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getVariancesInParameterEstimates}
\alias{getVariancesInParameterEstimates}
\title{getVariancesInParameterEstimates}
\usage{
getVariancesInParameterEstimates(mxObject, parameterEstimates)
}
\arguments{
\item{mxObject}{mxObject}

\item{parameterEstimates}{parameter estimates from regularized ctsem}
}
\description{
computes the variances from the parameterEstimates matrix
}
\details{
NOTE: Function located in file regCtsem.R
}
