% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIST.R
\name{GIST}
\alias{GIST}
\title{GIST}
\usage{
GIST(
  gradientModel,
  cppmodel,
  startingValues,
  objective,
  lambda,
  adaptiveLassoWeights,
  regularizedParameters,
  eta = 1.5,
  sig = 0.2,
  initialStepsize = 1,
  stepsizeMin = 0,
  stepsizeMax = 999999999,
  GISTLinesearchCriterion = "monotone",
  GISTNonMonotoneNBack = 5,
  maxIter_out = 100,
  maxIter_in = 100,
  break_outer = 1e-08,
  differenceApprox,
  verbose = 0,
  silent = FALSE
)
}
\arguments{
\item{cppmodel}{model of type cpptsemmodel}

\item{startingValues}{named vector with starting values}

\item{objective}{"ML" for maximum likelihood SEM, "Kalman" for Kalman filter}

\item{lambda}{penalty value}

\item{adaptiveLassoWeights}{named vector with adaptive lasso weights}

\item{regularizedParameters}{named vector of regularized parameters}

\item{eta}{if the current step size fails, eta will decrease the step size. Must be > 1}

\item{sig}{GIST: sigma value in Gong et al. (2013). Sigma controls the inner stopping criterion and must be in (0,1). Generally, a larger sigma enforce a steeper decrease in the regularized likelihood while a smaller sigma will result in faster acceptance of the inner iteration.}

\item{initialStepsize}{initial stepsize to be tried in the outer iteration}

\item{stepsizeMin}{Minimal acceptable step size. Must be > 0. A larger number corresponds to a smaller step from one to the next iteration. All step sizes will be computed as described by Gong et al. (2013)}

\item{stepsizeMax}{Maximal acceptable step size. Must be > stepsizeMin. A larger number corresponds to a smaller step from one to the next iteration. All step sizes will be computed as described by Gong et al. (2013)}

\item{GISTLinesearchCriterion}{criterion for accepting a step. Possible are 'monotone' which enforces a monotone decrease in the objective function or 'non-monotone' which also accepts some increase.}

\item{GISTNonMonotoneNBack}{in case of non-monotone line search: Number of preceding regM2LL values to consider}

\item{maxIter_out}{maximal number of outer iterations}

\item{maxIter_in}{maximal number of inner iterations}

\item{break_outer}{stopping criterion for the outer iteration.}

\item{verbose}{set to 1 to print additional information and plot the convergence}
}
\description{
General Iterative Shrinkage and Thresholding Algorithm based on Gong, P., Zhang, C., Lu, Z., Huang, J., & Ye, J. (2013). A General Iterative Shrinkage and Thresholding Algorithm for Non-convex Regularized Optimization Problems. In S. Dasgupta & D. McAllester (Eds.), Proceedings of Machine Learning Research (PMLR; Vol. 28, Issue 2, pp. 37--45). PMLR. http://proceedings.mlr.press
}
\details{
GIST minimizes a function of form f(theta) = l(theta) + g(theta), where l is the likelihood and g is a penalty function. Various penalties are supported, however currently only lasso and adaptive lasso are implemented.

NOTE: Function located in file GIST.R
}
