% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{gradCpptsem}
\alias{gradCpptsem}
\title{gradCpptsem}
\usage{
gradCpptsem(
  parameterValues,
  cpptsemObject,
  objective,
  free = labeledFree(parameterValues),
  failureReturns = NA
)
}
\arguments{
\item{parameterValues}{vector with labeled parameter values}

\item{cpptsemObject}{model of type cpptsem}

\item{objective}{ml or Kalman}

\item{free}{vector of same length as cpptsemObject$getParameterValues, where for each element it is specified if the parameter is freely estimated (workaround for fixing parameters)}

\item{failureReturns}{value which is returned if gradCpptsem fails}
}
\description{
gradCpptsem will try to compute gradients with decreasing precision starting from the default in OpenMx. Allows for setting some parameters to fixed (difference to exact_getCppGradients)
}
\keyword{internal}
