% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{individualGradient}
\alias{individualGradient}
\title{individualGradient}
\usage{
individualGradient(regCtsemObject, parameterValues, i, eps = 1e-04)
}
\arguments{
\item{regCtsemObject}{regularized model}

\item{parameterValues}{values for the parameters}

\item{i}{indicator for person(s)}

\item{eps}{value for numerical approximation of gradients}
}
\description{
compute the gradients for a singe person
}
