% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{individualHessianVectorProduct}
\alias{individualHessianVectorProduct}
\title{individualHessianVectorProduct}
\usage{
individualHessianVectorProduct(
  regCtsemObject,
  parameterValues,
  b,
  i,
  eps1 = 1e-04,
  eps2 = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{regCtsemObject}{regularized model}

\item{parameterValues}{values for the parameters}

\item{b}{the vector b which is multiplied with the Hessian}

\item{i}{indicator for person(s)}

\item{eps}{value for numerical approximation of gradients}
}
\description{
Computes H(theta)*b, where H(theta) is the individual Hessian at parameter vector theta and b is a vector which is
multiplied with H(theta). The procedure is based on Equation 1 in Pearlmutter, B. A. (1994). Fast Exact Multiplication by the Hessian.
Neural Computation, 6(1), 147–160. https://doi.org/10.1162/neco.1994.6.1.147
This is the most simple approximation and may be instable,
we should consider implementing the more sophisticated version by Pearlmutter, B. A. (1994).
}
