% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{individualM2LL}
\alias{individualM2LL}
\title{individualM2LL}
\usage{
individualM2LL(regCtsemObject, parameterValues, i)
}
\arguments{
\item{regCtsemObject}{regularized model}

\item{parameterValues}{values for the parameters}

\item{i}{indicator for person(s)}
}
\description{
compute the -2 log Likelihood for a singe person
}
