% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{individualM2LLKalman}
\alias{individualM2LLKalman}
\title{individualM2LLKalman}
\usage{
individualM2LLKalman(cpptsemObject, i)
}
\arguments{
\item{cpptsemObject}{cpptsemObject of class Rcpp_cpptsemKalmanModel}

\item{i}{index of individual}
}
\description{
Compute the -2log-Likelihood for a single individual
}
