% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{individualM2LLML}
\alias{individualM2LLML}
\title{individualM2LLML}
\usage{
individualM2LLML(cpptsemObject, wideData, i)
}
\arguments{
\item{cpptsemObject}{cpptsemObject of class Rcpp_cpptsemRAMmodel}

\item{wideData}{dataset in wide format}

\item{i}{index for person}
}
\description{
Compute the -2log-Likelihood for a single individual
}
