% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{ipc}
\alias{ipc}
\title{ipc}
\usage{
ipc(
  regCtsemObject,
  breakCondition = 1e-06,
  maxIterations = 100,
  gradientStepSize = 1e-04
)
}
\arguments{
\item{regCtsemObject}{regularized model}

\item{breakCondition}{break condition for the coordinate descent algorithm}

\item{maxIterations}{maximal number of iterations for the coordinate descent algorithm}

\item{gradientStepSize}{step size used for the gradient approximation}
}
\description{
ipc
}
