% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{ipcGlmnet}
\alias{ipcGlmnet}
\title{ipcGlmnet}
\usage{
ipcGlmnet(
  parameterValues,
  fitfunction,
  data,
  regIndicators,
  lambda,
  hessian = NULL,
  breakCondition = 1e-06,
  maxIterations = 100,
  gradientStepSize = 1e-04,
  ...
)
}
\arguments{
\item{parameterValues}{labeled vector with parameter values}

\item{fitfunction}{function which takes the data of a single person (vector), the parameter values and the additional arguments specified with ... as input and returns the fit value for this person}

\item{data}{data set with persons in rows and variables in columns}

\item{regIndicators}{vector with labels of regularized parameters}

\item{lambda}{size of the penalty value}

\item{hessian}{matrix with hessian}

\item{breakCondition}{break condition for the coordinate descent algorithm}

\item{maxIterations}{maximal number of iterations for the coordinate descent algorithm}

\item{gradientStepSize}{step size used for the gradient approximation}

\item{...}{additional arguments passed to fitfunction}
}
\description{
ipcGlmnet
}
