% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterateOverCVFolds.R
\name{iterateOverCVFolds}
\alias{iterateOverCVFolds}
\title{iterateOverCVFolds}
\usage{
iterateOverCVFolds(argsIn, objective = "ML", optimization)
}
\arguments{
\item{argsIn}{list of parameters passed to regCtsem}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{optimization}{type of optimization. Either exact or approx}
}
\description{
computes results for automativ cross-validation
}
