% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kalmanFit}
\alias{kalmanFit}
\title{kalmanFit}
\usage{
kalmanFit(
  update,
  sampleSize,
  Tpoints,
  nlatent,
  nmanifest,
  kalmanData,
  latentScores,
  predictedManifestValues,
  discreteTimeParameterNames,
  T0MEANSValues,
  T0VARValues,
  discreteDRIFTUnique,
  discreteCINTUnique,
  discreteTRAITUnique,
  discreteDIFFUSIONUnique,
  LAMBDAValues,
  MANIFESTMEANSValues,
  MANIFESTVARValues
)
}
\arguments{
\item{update}{boolean}

\item{sampleSize}{number of persons}

\item{Tpoints}{number of time points}

\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{kalmanData}{data for Kalman filter}

\item{latentScores}{matrix with latent scores}

\item{predictedManifestValues}{matrix with manifest predictions}

\item{discreteTimeParameterNames}{names of discrete time paramters}

\item{T0MEANSValues}{vector with initial means}

\item{T0VARValues}{matrix with initial variances}

\item{discreteDRIFTUnique}{list with discrete drift values}

\item{discreteCINTUnique}{list with discrete continuous time intercepts}

\item{discreteTRAITUnique}{list with discrete traits}

\item{discreteDIFFUSIONUnique}{list with discrete diffusions}

\item{LAMBDAValues}{matrix with loadings}

\item{MANIFESTMEANSValues}{manifest means}

\item{MANIFESTVARValues}{manifest variance}
}
\description{
Performs the prediction and updating step for the Kalman Filter
The implementation closely follows that of Driver, C. C., Oud, J. H. L., & Voelkle, M. C. (2017). Continuous Time Structural Equation Modelling With R Package ctsem. Journal of Statistical Software, 77(5), 1–36. https://doi.org/10.18637/jss.v077.i05
}
\keyword{internal}
