% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{labeledFree}
\alias{labeledFree}
\title{labeledFree}
\usage{
labeledFree(x)
}
\arguments{
\item{x}{vector with labeled values}
}
\description{
small helper function. Returns vector with TRUE of length x with same labels as x
}
\keyword{internal}
