% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{optimizeCpptsem}
\alias{optimizeCpptsem}
\title{optimizeCpptsem}
\usage{
optimizeCpptsem(cpptsemObject, free = "all", ...)
}
\arguments{
\item{cpptsemObject}{model of type cpptsem}

\item{...}{additional arguments passed to Rsonlp::solnp}
}
\description{
generic optimization with Rslonp
}
