% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regCtsem.R
\name{plot.regCtsem}
\alias{plot.regCtsem}
\title{plot.regCtsem}
\usage{
\method{plot}{regCtsem}(
  regCtsemObject,
  what = "drift",
  criterion = "BIC",
  xlab = "auto",
  ylab = "auto",
  ...
)
}
\arguments{
\item{regCtsemObject}{fitted regCtsem object}

\item{what}{what should be plotted? Possbile are: 'drift', 'parameters', 'fit' or labels of specific parameters}

\item{criterion}{vector with labels of criteria which should be plotted when what = 'fit'}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{additional parameters for plot or matplot (e.g., lty, col, ...)}
}
\description{
plot.regCtsem
}
\author{
Jannik Orzek
}
