% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regCtsem.R
\name{plot.regCtsemCV}
\alias{plot.regCtsemCV}
\title{plot.regCtsemCV}
\usage{
\method{plot}{regCtsemCV}(x, y = NULL, ..., xlab = "auto", ylab = "auto")
}
\arguments{
\item{x}{fitted regCtsem object}

\item{y}{NULL}

\item{...}{additional parameters for plot or matplot (e.g., lty, col, ...)}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}
}
\description{
plot.regCtsemCV
}
\author{
Jannik Orzek
}
