% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareDiscreteElementNames}
\alias{prepareDiscreteElementNames}
\title{prepareDiscreteElementNames}
\usage{
prepareDiscreteElementNames(ctMatrices, Tpoints, dT)
}
\arguments{
\item{ctMatrices}{continuous time matrices}

\item{Tpoints}{number of time points}

\item{dT}{time differences}
}
\description{
prepare the names of the discrete parameters (e.g. discrete time drift)
}
\keyword{internal}
