% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareDiscreteElements}
\alias{prepareDiscreteElements}
\title{prepareDiscreteElements}
\usage{
prepareDiscreteElements(mxObject, ctMatrices, nlatent, nmanifest, Tpoints, dT)
}
\arguments{
\item{mxObject}{Object of type mxObject}

\item{ctMatrices}{continuous time matrices}

\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{Tpoints}{number of time points}

\item{dT}{time differences}
}
\description{
prepare matrices for discrete time elements
}
\keyword{internal}
