% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareFMatrix}
\alias{prepareFMatrix}
\title{prepareFMatrix}
\usage{
prepareFMatrix(nlatent, nmanifest, Tpoints)
}
\arguments{
\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{Tpoints}{number of time points}
}
\description{
prepare filter matrix for RAM model
}
\keyword{internal}
