% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareKalmanData}
\alias{prepareKalmanData}
\title{prepareKalmanData}
\usage{
prepareKalmanData(dataset, nlatent, nmanifest, dtIndicators, Tpoints)
}
\arguments{
\item{dataset}{data set}

\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{dtIndicators}{indicators for discrete time parameters}

\item{Tpoints}{number of time points}
}
\description{
prepare data for Kalman model
}
\keyword{internal}
