% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareKalmanMatrices}
\alias{prepareKalmanMatrices}
\title{prepareKalmanMatrices}
\usage{
prepareKalmanMatrices(nlatent, nmanifest, Tpoints, sampleSize)
}
\arguments{
\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{Tpoints}{number of time points}

\item{sampleSize}{number of persons}
}
\description{
prepare matrices for Kalman model
}
\keyword{internal}
