% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareMMatrix}
\alias{prepareMMatrix}
\title{prepareMMatrix}
\usage{
prepareMMatrix(
  mxObject,
  ctMatrices,
  nlatent,
  nmanifest,
  Tpoints,
  dT,
  stationaryT0MEANS
)
}
\arguments{
\item{mxObject}{Object of type mxObject}

\item{ctMatrices}{continuous time matrices}

\item{nlatent}{number of latent variables}

\item{nmanifest}{number of manifest variables}

\item{Tpoints}{number of time points}

\item{dT}{time intervals}

\item{stationaryT0MEANS}{boolean: are Means stationary?}
}
\description{
prepare matrix with means for RAM model
}
\keyword{internal}
