% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{prepareRAMData}
\alias{prepareRAMData}
\title{prepareRAMData}
\usage{
prepareRAMData(dataset, individualMissingPatternID, uniqueMissingPatterns)
}
\arguments{
\item{dataset}{dataset}

\item{individualMissingPatternID}{IDs for individual missingness pattern}

\item{uniqueMissingPatterns}{unique missingness patterns}
}
\description{
prepare data for RAM model
}
