% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{print.regCtsemCVSummary}
\alias{print.regCtsemCVSummary}
\title{print summary of regCtsemCV}
\usage{
\method{print}{regCtsemCVSummary}(x, ...)
}
\arguments{
\item{x}{object of class regCtsemCV}

\item{...}{not used}
}
\value{
nothing
}
\description{
print summary of regCtsemCV
}
