% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{print.regCtsemSummary}
\alias{print.regCtsemSummary}
\title{print summary of regCtsem}
\usage{
\method{print}{regCtsemSummary}(x, ...)
}
\arguments{
\item{x}{object of class regCtsem}

\item{...}{not used}
}
\value{
nothing
}
\description{
print summary of regCtsem
}
