% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profileLikelihood}
\alias{profileLikelihood}
\title{profileLikelihood}
\usage{
profileLikelihood(
  cpptsemObject,
  parameter,
  values,
  lambda,
  startingValues = cpptsemObject$getParameterValues(),
  adaptiveLassoWeights,
  N,
  regIndicators,
  targetVector,
  epsilon = 1e-04,
  controlRsolnp = regCtsem::controlRsolnp()
)
}
\arguments{
\item{cpptsemObject}{model of type cpptsem}

\item{parameter}{label of the parameter for which the profile likelihood should be computed}

\item{values}{values for the parameter at which the profile likelihood will be computed}

\item{lambda}{penalty value}

\item{startingValues}{starting values}
}
\description{
computes the profile likelihood for a selected parameter
}
