% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profileLikelihoodFromCpptsem}
\alias{profileLikelihoodFromCpptsem}
\title{profileLikelihoodFromCpptsem}
\usage{
profileLikelihoodFromCpptsem(cpptsemObject, parameter, values)
}
\arguments{
\item{parameter}{label of the parameter for which the profile likelihood should be computed}

\item{values}{values for the parameter at which the profile likelihood will be computed}

\item{regCtsemObject}{model of type regCtsem}

\item{lambda}{penalty value}
}
\description{
computes the profile likelihood for a selected parameter based on a cpptsem
}
\examples{
\dontrun{
library(regCtsem)

CINT = matrix(c(0,0), nrow = 2, ncol = 1)

stationary <- c('T0TRAITEFFECT','T0TIPREDEFFECT')

## ctsem model without trait
AnomAuthmodel1 <- ctModel(LAMBDA = matrix(c(1, 0, 0, 1), nrow = 2, ncol = 2),
                          Tpoints = 5, n.latent = 2, n.manifest = 2,
                          MANIFESTVAR=diag(0, 2),
                          TRAITVAR = NULL,
                          CINT = CINT)
AnomAuthfit1 <- ctFit(AnomAuth, AnomAuthmodel1, useOptimizer = TRUE, stationary = stationary)

## with cpptsem
cpptsemModel <- cpptsemFromCtsem(ctsemModel = AnomAuthfit1, wideData = AnomAuth)
pl <- profileLikelihoodFromCpptsem(cpptsemObject = cpptsemModel, parameter = "drift_eta1_eta2", values = seq(-1,1,.2))
}
}
