% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profileLikelihoodFromMx}
\alias{profileLikelihoodFromMx}
\title{profileLikelihoodFromMx}
\usage{
profileLikelihoodFromMx(
  mxModelObject,
  parameter,
  values,
  startingValues = omxGetParameters(mxModelObject),
  tryHard = FALSE
)
}
\arguments{
\item{mxModelObject}{model from OpenMx}

\item{parameter}{label of the parameter for which the profile likelihood should be computed}

\item{values}{values for the parameter at which the profile likelihood will be computed}

\item{startingValues}{starting values}
}
\description{
computes the profile likelihood for a selected parameter based on an mxModel
}
