% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regIndicatorsFromNameToMatrix.R
\name{regIndicatorsFromNameToMatrix}
\alias{regIndicatorsFromNameToMatrix}
\title{regIndicatorsFromNameToMatrix}
\usage{
regIndicatorsFromNameToMatrix(mxObject, regOn, regIndicators)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{regOn}{string specifying which matrix should be regularized. Currently only supports DRIFT}

\item{regIndicators}{matrix with ones and zeros specifying which parameters in regOn should be regularized. Must be of same size as the regularized matrix. 1 = regularized, 0 = not regularized. Alternatively, labels for the regularized parameters can be used (e.g. drift_eta1_eta2)}
}
\description{
transforms string indicated regIndicators to matrix indicated regIndicators
}
