% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{restore}
\alias{restore}
\title{restore}
\usage{
restore(regCtsemObject)
}
\arguments{
\item{regCtsemObject}{object of type regCtsem}
}
\description{
restore the cpptsem object in a regCtsem object.
}
\details{
After saving and loading a regCtsem object, the underlying C++ model will be lost. This function resores that model
}
