% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximateLOOCV.R
\name{returnIndividualFit}
\alias{returnIndividualFit}
\title{returnIndividualFit}
\usage{
returnIndividualFit(individualData, parameterValues, fullData, cpptsemObject)
}
\arguments{
\item{individualData}{vector with data of one individual}

\item{parameterValues}{parameter values}

\item{fullData}{matrix with entire data set}

\item{cpptsemObject}{object created with cpptsemFromCtsem}
}
\description{
This function returns the fit of the model for the vector individualData. It is currently not very efficient
}
