% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{ridgeRAMRegM2LLCpptsem}
\alias{ridgeRAMRegM2LLCpptsem}
\title{ridgeRAMRegM2LLCpptsem}
\usage{
ridgeRAMRegM2LLCpptsem(
  parameters,
  cpptsemmodel,
  adaptiveLassoWeights,
  N,
  lambda_,
  regIndicators,
  targetVector,
  epsilon,
  objective,
  failureReturns
)
}
\arguments{
\item{parameters}{parameter values}

\item{cpptsemmodel}{model returned from cpptsem}

\item{adaptiveLassoWeights}{vector with weights of the adaptive lasso}

\item{N}{sample size}

\item{lambda_}{tuning parameter lambda}

\item{regIndicators}{string vector with names of regularized parameters}

\item{targetVector}{named vector with values towards which the parameters are regularized}

\item{epsilon}{NOT USED; Only required for the optimizer to call the function}

\item{objective}{ML or Kalman}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}
}
\description{
approximates the regularized likelihood function using cpptsem and Full Information Maximum Likelihood
}
\author{
Jannik Orzek
}
\keyword{internal}
