% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{separateFitAndParameters}
\alias{separateFitAndParameters}
\title{separateFitAndParameters}
\usage{
separateFitAndParameters(regCtsemObject)
}
\arguments{
\item{regCtsemObject}{Object from exact_regCtsem or approx_regCtsem}
}
\description{
separates fit results from parameter estimates
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
\keyword{internal}
