% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{setStartingValuesFromApprox}
\alias{setStartingValuesFromApprox}
\title{setStartingValuesFromApprox}
\usage{
setStartingValuesFromApprox(approx_regModel, mxObject, lambda)
}
\arguments{
\item{approx_regModel}{fitted regCtsem with optimization = "approx" and without automatic cross-validation}

\item{mxObject}{Fitted object of class MxObject}

\item{lambda}{single lambda}
}
\description{
set the parameter values of mxObject to the values in approx_regModel for the specified lambda
}
\details{
NOTE: Function located in file exact_optimization.R
}
\author{
Jannik Orzek
}
