% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shinyfy}
\alias{shinyfy}
\title{shinyfy}
\usage{
shinyfy(regCtsemObject, runShiny = TRUE, figHeight = 600)
}
\arguments{
\item{regCtsemObject}{regCtsemObject}

\item{runShiny}{boolean: if TRUE, the app is started directly. If false, the code to generate the app is returned}

\item{figHeight}{height of the figure}
}
\description{
creates a shiny application from a regCtsemObject
}
