% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showParameters}
\alias{showParameters}
\title{showParameters}
\usage{
showParameters(ctsemObject)
}
\arguments{
\item{ctsemObject}{Fitted object of class ctsemFit}
}
\description{
shows the parameters of a model fitted with ctFit.
Importantly, the untransformed paraemters are returned.
For instance, the variance-covariance matrices are implemented in log-Cholesky form
(see Pinheiro, J. C., & Bates, D. M. (1996). Unconstrained parametrizations for variance-covariance matrices. Statistics and Computing, 6(3), 289–296. https://doi.org/10.1007/BF00140873).
}
\examples{
library(regCtsem)

# The following example is taken from ?ctFit:
data(AnomAuth)
AnomAuthmodel <- ctModel(LAMBDA = matrix(c(1, 0, 0, 1), nrow = 2, ncol = 2),
                         Tpoints = 5, n.latent = 2, n.manifest = 2, MANIFESTVAR=diag(0, 2), TRAITVAR = NULL)
AnomAuthfit <- ctFit(AnomAuth, AnomAuthmodel)

showParameters(AnomAuthfit)
}
