% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsem}
\alias{summary.regCtsem}
\title{summary.regCtsem}
\usage{
\method{summary}{regCtsem}(regCtsemObject, criterion = NULL)
}
\arguments{
\item{criterion}{select a criterion. Possible are AIC, BIC, cvM2LL}

\item{returnSummary}{Boolean: should the summary results be returned}
}
\description{
summary.regCtsem
}
\author{
Jannik Orzek
}
