% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsemCV}
\alias{summary.regCtsemCV}
\title{summary.regCtsemCV}
\usage{
\method{summary}{regCtsemCV}(regCtsemObject)
}
\arguments{
\item{regCtsemObject}{Object of type regCtsemCV}
}
\description{
summary.regCtsemCV
}
\author{
Jannik Orzek
}
