% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsemMultiSubject}
\alias{summary.regCtsemMultiSubject}
\title{summary.regCtsemMultiSubject}
\usage{
\method{summary}{regCtsemMultiSubject}(regCtsemObject, criterion = NULL)
}
\arguments{
\item{regCtsemObject}{Object of type regCtsemMultiSubject}

\item{criterion}{select a criterion. Possible are AIC or BIC}
}
\description{
summary.regCtsemMultiSubject
}
\author{
Jannik Orzek
}
