% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_optimization.R
\name{approx_getFitIndices}
\alias{approx_getFitIndices}
\title{approx_getFitIndices}
\usage{
approx_getFitIndices(
  m2LL,
  regM2LL,
  lambda,
  parameterValues,
  targetVector,
  sampleSize,
  regIndicators,
  returnFitIndices,
  zeroThresh
)
}
\arguments{
\item{m2LL}{-2 log Likelihood}

\item{regM2LL}{regularized -2 log Likelihood}

\item{lambda}{tuning parameter lambda}

\item{parameterValues}{current parameter values}

\item{targetVector}{named vector with values towards which the parameters are regularized (Standard is regularization towards zero)}

\item{sampleSize}{sample size}

\item{regIndicators}{matrix with ones and zeros specifying which parameters in regOn should be regularized. Must be of same size as the regularized matrix. 1 = regularized, 0 = not regularized. Alternatively, labels for the regularized parameters can be used (e.g. drift_eta1_eta2)}

\item{returnFitIndices}{Boolean: should fit indices be returned?}

\item{zeroThresh}{threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx}
}
\description{
computes fit indices
}
\details{
NOTE: Function located in file approx_optimization.R
}
\author{
Jannik Orzek
}
