% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{approx_gradCpptsem}
\alias{approx_gradCpptsem}
\title{approx_gradCpptsem}
\usage{
approx_gradCpptsem(
  parameters,
  cpptsemmodel,
  adaptiveLassoWeights,
  N,
  lambda,
  regIndicators,
  targetVector,
  epsilon,
  objective,
  failureReturns
)
}
\arguments{
\item{parameters}{parameter values}

\item{cpptsemmodel}{Model of type cpptsem}

\item{adaptiveLassoWeights}{vector with weights of the adaptive lasso}

\item{N}{sample size}

\item{lambda}{tuning parameter lambda}

\item{regIndicators}{string vector with names of regularized parameters}

\item{targetVector}{named vector with values towards which the parameters are regularized}

\item{epsilon}{tuning parameter for epsL1 approximation}

\item{objective}{ML or Kalman}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}
}
\description{
computes gradients for an approximate optimization of regularized ctsem based on cpptsem
}
\author{
Jannik Orzek
}
