% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{approx_RAMM2LLCpptsem}
\alias{approx_RAMM2LLCpptsem}
\title{approx_RAMM2LLCpptsem}
\usage{
approx_RAMM2LLCpptsem(parameters, cpptsemmodel, failureReturns)
}
\arguments{
\item{parameters}{parameter values}

\item{cpptsemmodel}{model from cpptsem}

\item{failureReturns}{value which is returned if regM2LLCpptsem or gradCpptsem fails}
}
\description{
computes fit for RAM model
}
\author{
Jannik Orzek
}
\keyword{internal}
