% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_helper.R
\name{checkFI}
\alias{checkFI}
\title{checkFI}
\usage{
checkFI(mxObject, regCtsemObject, cvModel = NULL, threshold, testIC)
}
\arguments{
\item{mxObject}{object from openmx}

\item{regCtsemObject}{object from regCtsem}

\item{cvModel}{optional cross-validation model}

\item{threshold}{how close to zero should differences be to be treated as zero?}

\item{testIC}{should information criteria be tested?}
}
\description{
used in testthat to check the computation of fit indices of regCtsem
}
\keyword{internal}
