% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeDRIFTHASH}
\alias{computeDRIFTHASH}
\title{Computes the computeDRIFTHASH given the drift. DRIFTHASH = DRIFT otimes I + I otimes DRIFT.}
\usage{
computeDRIFTHASH(DRIFTValues)
}
\arguments{
\item{DRIFTValues}{matrix with drift values}
}
\value{
matrix with drift hash
}
\description{
Computes the computeDRIFTHASH given the drift. DRIFTHASH = DRIFT otimes I + I otimes DRIFT.
}
\keyword{internal}
