% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlDocumentation.R
\name{controlApprox}
\alias{controlApprox}
\title{controlApprox}
\usage{
controlApprox()
}
\arguments{
\item{epsilon}{epsilon is used to transform the non-differentiable lasso penalty to a differentiable one if optimization = approx}

\item{zeroThresh}{threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx}
}
\description{
The following arguments can be used to adjust the approximate optimization
}
