% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlDocumentation.R
\name{controlApprox}
\alias{controlApprox}
\title{controlApprox}
\usage{
controlApprox(
  forceCpptsem = FALSE,
  epsilon = 0.001,
  zeroThresh = 0.04,
  nMultistart = 5,
  controlApproxOptimizer = controlRsolnp(control = list(outer.iter = 500, trace = 0))
)
}
\arguments{
\item{forceCpptsem}{should cpptsem be enforced even if results differ from ctsem? Sometimes differences between cpptsem and ctsem can result from problems with numerical precision which will lead to the matrix exponential of RcppArmadillo differing from the OpenMx matrix exponential. If you want to ensure the faster optimization, set to TRUE. See vignette("MatrixExponential", package = "regCtsem") for more details}

\item{epsilon}{epsilon is used to transform the non-differentiable lasso penalty to a differentiable one if optimization = approx}

\item{zeroThresh}{threshold below which parameters will be evaluated as == 0 in lasso regularization if optimization = approx}

\item{nMultistart}{controls how many different starting values are tried when estimating lambda_max}

\item{controlApproxOptimizer}{settings passed to the optimizer in approximate optimization. Currently, Rsolnp and optimx are supported. See ?controlOptimx and ?controlSolnp for details on the lists passed to controlApprox}
}
\description{
The following arguments can be used to adjust the approximate optimization
}
