% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{createCVFoldsAndModels}
\alias{createCVFoldsAndModels}
\title{createCVFoldsAndModels}
\usage{
createCVFoldsAndModels(mxObject, dataset = NULL, k)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject.}

\item{dataset}{only required if objective = "Kalman" and ctsemObject ist of type ctsemInit. Please provide a data set in wide format compatible to ctsemOMX}

\item{k}{number of cross-validation folds if autoCV = TRUE (k-fold cross-validation)}
}
\description{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
