% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{createFolds}
\alias{createFolds}
\title{createFolds}
\usage{
createFolds(mxObject, dataset = NULL, k)
}
\arguments{
\item{mxObject}{Object of class mxModel}

\item{dataset}{data set}

\item{k}{number of cross-validation folds (k-fold cross-validation)}
}
\description{
created folds for automatic cross-validation
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
