% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.R
\name{createFolds}
\alias{createFolds}
\title{createFolds}
\usage{
createFolds(mxObject, dataset = NULL, k)
}
\arguments{
\item{mxObject}{Object of class mxModel}

\item{dataset}{only required if objective = "Kalman" and ctsemObject ist of type ctsemInit. Please provide a data set in wide format compatible to ctsemOMX}

\item{k}{number of cross-validation folds (k-fold cross-validation)}
}
\description{
created folds for automatic cross-validation
}
\author{
Jannik Orzek
}
