% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_armijoLineSearch}
\alias{exact_armijoLineSearch}
\title{exact_armijoLineSearch}
\usage{
exact_armijoLineSearch(
  gradientModel,
  adaptiveLassoWeights,
  thetaNames,
  regIndicators,
  lambda,
  theta_kp1,
  m2LL_kp1,
  g_kp1,
  d,
  c1 = 1e-04,
  stepSize = 0.99,
  differenceApprox = "central",
  maxIter_line = 100
)
}
\arguments{
\item{gradientModel}{mxObject for computing the derivarive of the likelihood with respect to the parameter estimates}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{thetaNames}{names of the parameter estimates}

\item{regIndicators}{vector with names of parameters to regularize}

\item{lambda}{penalty value}

\item{theta_kp1}{parameter values of iteration k plus 1}

\item{m2LL_kp1}{-2 log likelihood of iteration k plus 1}

\item{g_kp1}{gradients of iteration k plus 1}

\item{d}{vector with updates to parameter estimates}

\item{c1}{tuning parameter for Armijo condition}

\item{stepSize}{Initial stepsize of the outer iteration (theta_{k+1} = theta_k + Stepsize \* Stepdirection)}

\item{differenceApprox}{which approximation for the gradients should be used? Recommended is central}

\item{maxIter_line}{maximal number of iterations for line search}
}
\description{
performs Armijo line search
}
\details{
NOTE: Function located in file GLMNET.R
}
