% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_getBFGS.R
\name{exact_getBFGS}
\alias{exact_getBFGS}
\title{exact_getBFGS}
\usage{
exact_getBFGS(theta_k, g_k, H_k, theta_kp1, g_kp1)
}
\arguments{
\item{theta_k}{Theta at iteration k}

\item{g_k}{Gradients of the likelihood function at iteration k}

\item{H_k}{Hessian of the likelihood function at iteration k}

\item{theta_kp1}{Theta at iteration k+1}

\item{g_kp1}{Gradients of the likelihood function at iteration k+1}
}
\description{
computes the BFGS Hessian approximation
}
