% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_getBFGS}
\alias{exact_getBFGS}
\title{exact_getBFGS}
\usage{
exact_getBFGS(
  theta_k,
  g_k,
  H_k,
  theta_kp1,
  g_kp1,
  cautious = TRUE,
  hessianEps = 0.001
)
}
\arguments{
\item{theta_k}{Theta at iteration k}

\item{g_k}{Gradients of the likelihood function at iteration k}

\item{H_k}{Hessian of the likelihood function at iteration k}

\item{theta_kp1}{Theta at iteration k+1}

\item{g_kp1}{Gradients of the likelihood function at iteration k+1}

\item{cautious}{boolean: should the update be skipped if it would result in a non positive definite Hessian?}

\item{hessianEps}{controls when the update of the Hessian approximation is skipped}
}
\description{
computes the BFGS Hessian approximation
}
\details{
NOTE: Function located in file GLMNET.R
}
