% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_getCVFit}
\alias{exact_getCVFit}
\title{exact_getCVFit}
\usage{
exact_getCVFit(
  objective,
  cvSampleCpptsemObject,
  parameterLabels,
  parameterValuesTable,
  lambdas,
  cvSample
)
}
\arguments{
\item{objective}{Kalman or mx}

\item{cvSampleCpptsemObject}{Fitted object of class cpptsem}

\item{parameterLabels}{labels of optimized parameters}

\item{parameterValuesTable}{table with parameter values+}

\item{lambdas}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}

\item{cvSample}{cross-validation sample. Has to be of type mxData}
}
\description{
computes cross-validation fit for exact optimization.
}
\details{
NOTE: Function located in file exact_optimization.R
}
\author{
Jannik Orzek
}
\keyword{internal}
