% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_getFitIndices.R
\name{exact_getFitIndices}
\alias{exact_getFitIndices}
\title{exact_getFitIndices}
\usage{
exact_getFitIndices(
  mxObject,
  parameterLabels,
  fitAndParameters,
  regValues,
  sampleSize
)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject}

\item{parameterLabels}{labels of optimized parameters}

\item{fitAndParameters}{table with fit and parameter values}

\item{regValues}{vector of penalty values (tuning parameter). E.g., seq(0,1,.01)}

\item{sampleSize}{sample size}
}
\description{
computes fit indices for optimization = "exact"
}
\author{
Jannik Orzek
}
