% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMNET.R
\name{exact_innerGLMNET}
\alias{exact_innerGLMNET}
\title{exact_innerGLMNET}
\usage{
exact_innerGLMNET(
  adaptiveLassoWeights,
  thetaNames,
  regIndicators,
  lambda,
  theta_kp1,
  g_kp1,
  H_kp1,
  maxIter_in,
  eps_in
)
}
\arguments{
\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{thetaNames}{Vector with names of theta-parameters}

\item{regIndicators}{Names of regularized parameters}

\item{lambda}{Penaltiy value}

\item{theta_kp1}{Theta at iteration k+1}

\item{g_kp1}{Gradients of the likelihood function at iteration k+1}

\item{H_kp1}{Hessian of the likelihood function at iteration k+1}

\item{maxIter_in}{Maximal number of iterations of the inner optimization algorithm}

\item{eps_in}{Stopping criterion for the inner iterations}
}
\description{
performs the inner optimization routine of GLMNET. exact_innerGLMNET returns a direction vector for the next step in the outer optimization routine.
}
\details{
NOTE: Function located in file GLMNET.R
}
