% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_tryStartingValues}
\alias{exact_tryStartingValues}
\title{exact_tryStartingValues}
\usage{
exact_tryStartingValues(
  gradientModel,
  cppmodel,
  objective,
  currentParameters,
  sparseParameters,
  regIndicators,
  newLambda,
  adaptiveLassoWeights,
  differenceApprox,
  numStartingValues,
  optimize,
  numOuter
)
}
\arguments{
\item{gradientModel}{Model used for computing gradients (from OpenMx)}

\item{cppmodel}{cpptsem object to compute gradients}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{currentParameters}{current parameter values}

\item{sparseParameters}{parameters of the sparsest model}

\item{regIndicators}{named vector of regularized parameters}

\item{newLambda}{new regValue}

\item{adaptiveLassoWeights}{named vector with adaptive lasso weights}

\item{differenceApprox}{difference approximation for gradients (detault should be central)}

\item{numStartingValues}{How many starting values should be tried?}

\item{optimize}{1 = optimization with optim, 2 = optimization with Rsolnp}

\item{numOuter}{number of outer iterations of optim or Rsolnp}
}
\description{
tries different starting values to find a good starting point for exact optimization
}
\details{
NOTE: Function located in file exact_optimization.R
}
