% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{exact_tryStartingValues}
\alias{exact_tryStartingValues}
\title{exact_tryStartingValues}
\usage{
exact_tryStartingValues(
  cpptsemObject,
  objective,
  currentParameters,
  sparseParameters,
  regIndicators,
  newLambda,
  adaptiveLassoWeights,
  numStartingValues,
  numOuter
)
}
\arguments{
\item{cpptsemObject}{cpptsem object to compute gradients}

\item{objective}{which objective should be used? Possible are "ML" (Maximum Likelihood) or "Kalman" (Kalman Filter)}

\item{currentParameters}{current parameter values}

\item{sparseParameters}{parameters of the sparsest model}

\item{regIndicators}{named vector of regularized parameters}

\item{newLambda}{new lambda}

\item{adaptiveLassoWeights}{named vector with adaptive lasso weights}

\item{numStartingValues}{How many starting values should be tried?}

\item{numOuter}{number of outer iterations of optim or Rsolnp}
}
\description{
tries different starting values to find a good starting point for exact optimization
}
\details{
NOTE: Function located in file exact_optimization.R
}
