% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{generateDRIFTPlot}
\alias{generateDRIFTPlot}
\title{generateDRIFTPlot}
\usage{
generateDRIFTPlot(
  model,
  ylab = "auto",
  xlab = "auto",
  skiptYlabComp,
  skiptXlabComp
)
}
\arguments{
\item{model}{model from regCtsem}

\item{ylab}{label for y-axis. auto will set to lambda}

\item{xlab}{label for x-axis. auto will set to drift}

\item{skiptYlabComp}{boolean: is ylab a string}

\item{skiptXlabComp}{boolean: is xlab a string}
}
\description{
generates plot of drift values
}
\details{
NOTE: Function located in file regCtsem.R
}
\keyword{internal}
