% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdaptiveLassoWeights.R
\name{getAdaptiveLassoWeights}
\alias{getAdaptiveLassoWeights}
\title{getAdaptiveLassoWeights}
\usage{
getAdaptiveLassoWeights(
  mxObject,
  penalty,
  adaptiveLassoWeights,
  standardizeDrift
)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{penalty}{type. Currently supported are lasso and ridge for optimization = approx and lasso for optimization = exact}

\item{adaptiveLassoWeights}{weights for the adaptive lasso.}

\item{standardizeDrift}{Boolean: Should Drift parameters be standardized automatically using T0VAR?}
}
\description{
Computes the adaptive lasso weights
}
\author{
Jannik Orzek
}
