% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFinalParameters.R
\name{getFinalParameters}
\alias{getFinalParameters}
\title{getFinalParameters}
\usage{
getFinalParameters(regCtsemObject, criterion = NULL)
}
\arguments{
\item{regCtsemObject}{fitted regularized continuous time model}

\item{criterion}{select a criterion. Possible are AIC, BIC, cvM2LL}
}
\description{
returns the final parameters for a
}
\author{
Jannik Orzek
}
