% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getFlatStdizer}
\alias{getFlatStdizer}
\title{getFlatStdizer}
\usage{
getFlatStdizer(T0VAR, driftLabels)
}
\arguments{
\item{T0VAR}{matrix with T0VAR values}

\item{driftLabels}{vector with drift names}
}
\description{
returns the standardizer for the standardized drift parameters if standardizeDrift = TRUE. Computes T0SD_predictor/T0SD_dependent
}
\details{
NOTE: Function located in file regCtsem.R
}
