% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regCtsem.R
\name{getMaxLambda}
\alias{getMaxLambda}
\title{getMaxLambda}
\usage{
getMaxLambda(mxObject, regIndicators, differenceApprox, adaptiveLassoWeights)
}
\arguments{
\item{mxObject}{Fitted object of class MxObject extracted from ctsemObject. Provide either ctsemObject or mxObject}

\item{regIndicators}{Labels of the regularized parameters (e.g. drift_eta1_eta2)}

\item{differenceApprox}{Which approximation should be used for calculating the gradients in the gradientModel. central is recommended}

\item{adaptiveLassoWeights}{weights for the adaptive lasso. If auto, defaults to the unregularized parameter estimates.}
}
\description{
computes an approximation of the lowest lambda which will set all regularized parameters to zero. This function is adapted from Murphy (2012) Machine learning: a probabilistic perspective. See p. 434 for more details.
}
\details{
NOTE: Function located in file regCtsem.R
}
\author{
Jannik Orzek
}
