% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarianceFromVarianceBase2.R
\name{getVarianceFromVarianceBase2}
\alias{getVarianceFromVarianceBase2}
\title{getVarianceFromVarianceBase2}
\usage{
getVarianceFromVarianceBase2(varianceBaseValues)
}
\arguments{
\item{varianceBaseValues}{values of varianceBase}
}
\description{
returns variance from varianceBase
}
\author{
Jannik Orzek
}
