% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpptsem.R
\name{optimizeCpptsem}
\alias{optimizeCpptsem}
\title{optimizeCpptsem}
\usage{
optimizeCpptsem(cpptsemObject, free = "all", nMultistart = 0, ...)
}
\arguments{
\item{cpptsemObject}{model of type cpptsem}

\item{free}{named logical vector indicating which of the parameters in cpptsemObject$getParameterValues() are free}

\item{nMultistart}{number of multi-start iterations}

\item{...}{additional arguments passed to Rsonlp::solnp}
}
\description{
generic optimization with Rslonp
}
