% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regCtsem.R
\name{plot.regCtsemCV}
\alias{plot.regCtsemCV}
\title{plot.regCtsemCV}
\usage{
\method{plot}{regCtsemCV}(regCtsemObject, xlab = "auto", ylab = "auto", ...)
}
\arguments{
\item{regCtsemObject}{fitted regCtsem object}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{additional parameters for plot or matplot (e.g., lty, col, ...)}
}
\description{
plot.regCtsemCV
}
\author{
Jannik Orzek
}
\keyword{internal}
