% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regCtsem.R
\name{plot.regCtsemMultiSubject}
\alias{plot.regCtsemMultiSubject}
\title{plot.regCtsemMultiSubject}
\usage{
\method{plot}{regCtsemMultiSubject}(
  regCtsemObject,
  what = "drift",
  groups = NULL,
  criterion = "BIC",
  xlab = "auto",
  ylab = "auto",
  ...
)
}
\arguments{
\item{regCtsemObject}{fitted regCtsem object}

\item{what}{what should be plotted? Possbile are: 'drift', 'parameters', 'fit'}

\item{groups}{For which groups should the parameters be plotted (numeric vector)?}

\item{criterion}{vector with labels of criteria which should be plotted when what = 'fit'}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{...}{additional parameters for plot or matplot (e.g., lty, col, ...)}
}
\description{
plot.regCtsemMultiSubject
}
\author{
Jannik Orzek
}
\keyword{internal}
