% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStartingValuesFromApprox.R
\name{setStartingValuesFromApprox}
\alias{setStartingValuesFromApprox}
\title{setStartingValuesFromApprox}
\usage{
setStartingValuesFromApprox(approx_regModel, mxObject, regValue)
}
\arguments{
\item{approx_regModel}{fitted regCtsem with optimization = "approx" and without automatic cross-validation}

\item{mxObject}{Fitted object of class MxObject}

\item{regValue}{single regValue}
}
\description{
set the parameter values of mxObject to the values in approx_regModel for the specified regValue
}
\author{
Jannik Orzek
}
