% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsem}
\alias{summary.regCtsem}
\title{summary.regCtsem}
\usage{
\method{summary}{regCtsem}(object, ..., criterion = NULL)
}
\arguments{
\item{object}{Object of type regCtsem}

\item{...}{NULL}

\item{criterion}{select a criterion. Possible are AIC, BIC, cvM2LL}
}
\description{
summary.regCtsem
}
\author{
Jannik Orzek
}
