% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsemCV}
\alias{summary.regCtsemCV}
\title{summary.regCtsemCV}
\usage{
\method{summary}{regCtsemCV}(object, ...)
}
\arguments{
\item{object}{Object of type regCtsemCV}

\item{...}{NULL}
}
\description{
summary.regCtsemCV
}
\author{
Jannik Orzek
}
