% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regCtsem.R
\name{summary.regCtsemMultiSubject}
\alias{summary.regCtsemMultiSubject}
\title{summary.regCtsemMultiSubject}
\usage{
\method{summary}{regCtsemMultiSubject}(object, ..., criterion = NULL)
}
\arguments{
\item{object}{Object of type regCtsemMultiSubject}

\item{...}{NULL}

\item{criterion}{select a criterion. Possible are AIC or BIC}
}
\description{
summary.regCtsemMultiSubject
}
\author{
Jannik Orzek
}
