% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_optimization.R
\name{tryApproxFirst}
\alias{tryApproxFirst}
\title{tryApproxFirst}
\usage{
tryApproxFirst(
  startingValues,
  returnAs,
  approxFirst,
  numStart,
  approxOpt,
  approxMaxIt,
  lambda,
  lambdas,
  gradientModelcpp,
  mxObject,
  regOn,
  regIndicators,
  adaptiveLassoWeights,
  objective,
  sparseParameters,
  extraTries
)
}
\description{
NOTE: Function located in file exact_optimization.R
}
\author{
Jannik Orzek
}
